### 2023-06-17 #5568 @danielskatz
👋 @devnums - your paper is not compiling. I think this is because you don't have an affiliation listed in the header in the paper.md file.  Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and note that you can click on the error above to find out more about it.

Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Your paper is also remarkably short. In addition to the example paper and the documentation about paper requirements, please see the other recently published papers for some examples.
