### 2023-06-19 #5568 @bmcfee
Thanks for the invite, but I'm unavailable for new assignments right now.
