+++ paper/paper.bib
@@ -53,50 +53,50 @@
 }
 
 @misc{tensorflow2015,
-title={ {TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
-url={https://www.tensorflow.org/},
-note={Software available from tensorflow.org},
-author={
-    Mart\'{i}n~Abadi and
-    Ashish~Agarwal and
-    Paul~Barham and
-    Eugene~Brevdo and
-    Zhifeng~Chen and
-    Craig~Citro and
-    Greg~S.~Corrado and
-    Andy~Davis and
-    Jeffrey~Dean and
-    Matthieu~Devin and
-    Sanjay~Ghemawat and
-    Ian~Goodfellow and
-    Andrew~Harp and
-    Geoffrey~Irving and
-    Michael~Isard and
-    Yangqing Jia and
-    Rafal~Jozefowicz and
-    Lukasz~Kaiser and
-    Manjunath~Kudlur and
-    Josh~Levenberg and
-    Dandelion~Man\'{e} and
-    Rajat~Monga and
-    Sherry~Moore and
-    Derek~Murray and
-    Chris~Olah and
-    Mike~Schuster and
-    Jonathon~Shlens and
-    Benoit~Steiner and
-    Ilya~Sutskever and
-    Kunal~Talwar and
-    Paul~Tucker and
-    Vincent~Vanhoucke and
-    Vijay~Vasudevan and
-    Fernanda~Vi\'{e}gas and
-    Oriol~Vinyals and
-    Pete~Warden and
-    Martin~Wattenberg and
-    Martin~Wicke and
-    Yuan~Yu and
-    Xiaoqiang~Zheng},
+  title={ {TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
+  url={https://www.tensorflow.org/},
+  note={Software available from tensorflow.org},
+  author={
+      Mart\'{i}n~Abadi and
+      Ashish~Agarwal and
+      Paul~Barham and
+      Eugene~Brevdo and
+      Zhifeng~Chen and
+      Craig~Citro and
+      Greg~S.~Corrado and
+      Andy~Davis and
+      Jeffrey~Dean and
+      Matthieu~Devin and
+      Sanjay~Ghemawat and
+      Ian~Goodfellow and
+      Andrew~Harp and
+      Geoffrey~Irving and
+      Michael~Isard and
+      Yangqing Jia and
+      Rafal~Jozefowicz and
+      Lukasz~Kaiser and
+      Manjunath~Kudlur and
+      Josh~Levenberg and
+      Dandelion~Man\'{e} and
+      Rajat~Monga and
+      Sherry~Moore and
+      Derek~Murray and
+      Chris~Olah and
+      Mike~Schuster and
+      Jonathon~Shlens and
+      Benoit~Steiner and
+      Ilya~Sutskever and
+      Kunal~Talwar and
+      Paul~Tucker and
+      Vincent~Vanhoucke and
+      Vijay~Vasudevan and
+      Fernanda~Vi\'{e}gas and
+      Oriol~Vinyals and
+      Pete~Warden and
+      Martin~Wattenberg and
+      Martin~Wicke and
+      Yuan~Yu and
+      Xiaoqiang~Zheng},
   year={2015},
 }
 
@@ -105,4 +105,24 @@ author={
   author={Chollet, Fran\c{c}ois and others},
   year={2015},
   url={https://keras.io}
+}
+
+@article{Makowski_NeuroKit2_A_Python_2021,
+  author = {
+      Makowski, Dominique and
+      Pham, Tam and
+      Lau, Zen J. and
+      Brammer, Jan C. and
+      Lespinasse, François and
+      Pham, Hung and
[…]
