### 2023-06-28 #5568 @britta-wstnr
Hi @peterakirk and @cudmore :wave:

Would you be willing to review the submission:
**pyheartlib: A Python package for processing electrocardiogram signals**
for [The Journal of Open Source Software (JOSS)](https://joss.theoj.org/)?

The review concerns the software [pyheartlib](https://github.com/devnums/pyheartlib) and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.05568/joss.05568/10.21105.joss.05568.pdf) about it. The review happens fully on GitHub.

You can learn more about the the review process and our conflict of interest policy in the reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

If you are available to review this submission, please let me know and I can add you as a reviewer. The review process will start (in a separate GitHub issue) once a sufficient number of reviewers have been found.

Thank you!
Britta
