### 2023-07-04 #5568 @cudmore
Yes, I am happy to review.

Robert

On Tue, Jul 4, 2023 at 1:11 AM Britta Westner ***@***.***>
wrote:

> ping @peterakirk <https://github.com/peterakirk> and @cudmore
> <https://github.com/cudmore>
> Kind reminder for my review invitation (see comment above
> <https://github.com/openjournals/joss-reviews/issues/5568#issuecomment-1611629045>)
> 🙏
>
