### 2023-07-21 #5568 @Gabrock94
> ping @Gabrock94 wave kind reminder for my review invitation (see comment [above](https://github.com/openjournals/joss-reviews/issues/5568#issuecomment-1625215495)) pray

HI @britta-wstnr , sorry I completely missed the first notification. I will work on it in the coming days.

Best,
Giulio
