### 2023-08-18 #5568 @TomDonoghue
Hey @britta-wstnr - thanks for the invite! Unfortunately, I have no practical experience with measuring cardiac signals to help me assess this package, and am also currently quite limited in my schedule, so I'm going to have to pass on this one!
