### 2023-09-13 devnums/pyheartlib#8 @LegrandNico
add more information to the contribution guidelines
The contribution guidelines could be detailed, adding for example instructions on how to contribute to the documentation and run the tests. Using a public platform for questions and help is probably preferable over email exchanges (also for the maintainer).

See this example of guidelines from another popular repository [here](https://github.com/blackjax-devs/blackjax/blob/main/CONTRIBUTING.md).

This issue is related to the review of the package for JOSS (see [here](https://github.com/openjournals/joss-reviews/issues/5792)).
