### 2023-09-13 devnums/pyheartlib#9 @LegrandNico
add legend to the paper's main figure
It is unclear to me what the label represents in the paper's main figure [see here](https://github.com/devnums/pyheartlib/blob/main/paper/mis.png). I think this one would benefit from some context about the processing pipeline and analysis goal.

For example, why are some of the R waves not labelled as True peaks when they seem normal?

What do false positive and negative mean exactly?

The prediction appears to be missing some of the R wave, is it expected or a problem?

This issue is related to the review of the package for JOSS (see [here](https://github.com/openjournals/joss-reviews/issues/5792)).
