### 2023-09-13 devnums/pyheartlib#10 @LegrandNico
add comparison with other libraries for peak detection
As the added value of the package is to facilitate the use of machine learning and deep learning methods to analyse ECG signal, it would be useful to see how this relate to other methods, especially methods that are simpler. I would suggest a comparison with the peak detection algorithms provided by [Neurokit2](https://github.com/neuropsychology/NeuroKit) and [SleepECG](https://github.com/cbrnr/sleepecg).

This issue is related to the review of the package for JOSS (see [here](https://github.com/openjournals/joss-reviews/issues/5792)).
