### 2023-09-19 9370700
style: edit to conform more to PEP 8 guidlines
