### 2023-09-19 d7b10b5
chore: improve example
