### 2023-09-19 7baf1ea
chore: add new file to separate from inference.py
