+++ examples/model/inference-plot.py
@@ -1,0 +1,172 @@
+#############################################################################
+# Copyright (c) 2023 Pyheartlib team. - All Rights Reserved                 #
+# Project repo: https://github.com/devnums/pyheartlib                       #
+# Contact: devnums.code@gmail.com                                           #
+#                                                                           #
+# This file is part of the Pyheartlib project.                              #
+# To see the complete LICENSE file visit:                                   #
+# https://github.com/devnums/pyheartlib/blob/main/LICENSE                   #
+#############################################################################
+
+
+# This file can be used to do inference using the trained example model
+import os
+
+import numpy as np
+import tensorflow as tf
+from tqdm import tqdm
+
+from pyheartlib.io import load_data
+
+cdir = os.path.dirname(os.path.abspath(__file__))
+os.chdir(cdir)
+print("Current directory changed to:\n", cdir)
+
+data_dir = "../../data"
+test_data = os.path.join(data_dir, "test.rpeak")
+mdl_checkpoint = os.path.join(os.getcwd(), "checkpoint/keras.exp")
+interval_value = 6
+opt = {1: "mis", 2: "correct", 3: "all"}
+to_plot = opt[3]
+rang = range(5240, 5246)
+
+# load trained model
+loaded_model = tf.keras.models.load_model(mdl_checkpoint)
+
+# load test data
+annotated_records_test, samples_info_test = load_data(test_data)
+print(len(samples_info_test))
+# plot
+import matplotlib.pyplot as plt  # noqa: E402
+
+from pyheartlib.data_rpeak import ECGSequence  # noqa: E402
+
+batch_size = 1
+labels_true = []
+labels_pred = []
+for i in tqdm(rang):  # range(round(len(samples_info_test)/batch_size)):
+    testseq = ECGSequence(
+        annotated_records_test,
+        [samples_info_test[i]],
+        binary=True,
+        batch_size=batch_size,
+        raw=True,
+        interval=interval_value,
+        shuffle=False,
+    )
+    samples = testseq[0]
+    labels_true = samples[1]
+    probs = loaded_model.predict(samples[0], verbose=0)
+    labels_pred = np.argmax(probs, axis=-1)
+    labels_true = np.array(labels_true)
+    labels_pred = np.array(labels_pred)
+    cond = not np.array_equal(labels_true, labels_pred)
+    if to_plot == "mis":
+        cond = not np.array_equal(labels_true, labels_pred)
+    elif to_plot == "correct":
+        cond = np.array_equal(labels_true, labels_pred)
+    elif to_plot == "all":
+        cond = True
+    if cond:
+        rec = samples_info_test[i][0]
+        st = samples_info_test[i][1]
+        en = samples_info_test[i][2]
+        sig = annotated_records_test[rec]["signal"][st:en]
+        # figure
+        plt.figure(figsize=(10, 5))
+        # plot signal
+        plt.plot(sig, linewidth=0.8, color="#118ed6", zorder=2)
+        # set axis range
+        plt.xlim(-10, 3610)
+        plt.ylim(-1.7, 2.0)
+        plt.xlabel("Samples")
+        plt.ylabel("Voltage(mV)")
+        # add grid lines
+        plt.grid(
+            which="major",
+            c="#bdbebf",
+            linestyle="dashed",
+            linewidth=0.5,
+            zorder=1,
+        )
+        plt.grid(
+            which="minor",
+            c="#bdbebf",
+            linestyle="dashed",
+            linewidth=0.3,
+            zorder=1,
+        )
+        plt.minorticks_on()
+        # marker size
+        mrkr_size = 30
+        for p in range(labels_true.shape[1]):
+            if labels_true[0][p] == 1:
+                plt.scatter(
[…]
