+++ examples/dataset/make_dataset_intra.py
@@ -1,10 +1,20 @@
+#############################################################################
+# Copyright (c) 2023 Pyheartlib team. - All Rights Reserved                 #
+# Project repo: https://github.com/devnums/pyheartlib                       #
+# Contact: devnums.code@gmail.com                                           #
+#                                                                           #
+# This file is part of the Pyheartlib project.                              #
+# To see the complete LICENSE file visit:                                   #
+# https://github.com/devnums/pyheartlib/blob/main/LICENSE                   #
+#############################################################################
+
+
 """
 This example shows how to create the datasets intra-patient.
 """
 
-from pyheartlib.data_beat import BeatData
 from pyheartlib.beat_info import BeatInfo
-
+from pyheartlib.data_beat import BeatData
 
 # Intra-patient
 # Lets create a BeatData object.
