+++ gen_license.sh
@@ -1,5 +1,16 @@
 #!/bin/bash
 
+#############################################################################
+# Copyright (c) 2023 Pyheartlib team. - All Rights Reserved                 #
+# Project repo: https://github.com/devnums/pyheartlib                       #
+# Contact: devnums.code@gmail.com                                           #
+#                                                                           #
+# This file is part of the Pyheartlib project.                              #
+# To see the complete LICENSE file visit:                                   #
+# https://github.com/devnums/pyheartlib/blob/main/LICENSE                   #
+#############################################################################
+
+
 poetry export -o requirements.textt --without-hashes
 pythonpath=$(poetry env info --path)
 pythonpath+=/bin/python
