+++ pyproject.toml
@@ -33,6 +33,9 @@ sphinx-rtd-theme = "^1.2.1"
 third-party-license-file-generator = "^2023.2.22"
 flake8 = "^6.0.0"
 python-semantic-release = "^7.33.5"
+pre-commit = "^3.4.0"
+isort = "^5.12.0"
+ruff = "^0.0.290"
 
 [tool.semantic_release]
 version_variable = "pyproject.toml:version" # version location
