### 2023-09-19 8702a83
ci: add pre-commit action
