+++ docs/examples/model/BEAT_DETECTION.rst
@@ -1,9 +1,9 @@
-==============
-Beat Detection
-==============
+================
+R-peak Detection
+================
 
-This example contains the code for training a deep learning model for the heartbeat detection task. As the task is not complex, it can be trained with high accuracy on a small amount of data.
-In this example, the electrocardiogram signals are segmented into ten seconds ecxerpts. Each excerpt corresponds to an annotation list of length 100 containing zeros and ones. A subsegment with an R-peak corresponds to one in the annotation list.
+This example contains the code for training a deep learning model for the R-peak detection task. As the task is not complex, it can be trained with high accuracy on a small amount of data.
+In this example, the electrocardiogram signals are segmented into ten seconds ecxerpts. Each excerpt corresponds to an annotation list of length 600 containing zeros and ones. A subsegment with an R-peak corresponds to one in the annotation list.
 
 The following steps prepare the data and train the model.
 
@@ -33,8 +33,6 @@ Results
 
 .. figure:: plots/mis.png
   :width: 600
-  :alt: Example: heartbeat detection using deep learning
+  :alt: Example: R-peak detection using deep learning
 
-  Heartbeat detection using deep learning.
-
-
+  R-peak detection using deep learning.
