+++ docs/examples/model/result.txt
@@ -1,12 +1,12 @@
               precision    recall  f1-score   support
 
-           0       0.99      1.00      0.99    862290
-           1       0.99      0.93      0.96    123310
+           0       0.99      1.00      1.00   5790290
+           1       0.92      0.72      0.81    123310
 
-    accuracy                           0.99    985600
-   macro avg       0.99      0.96      0.98    985600
-weighted avg       0.99      0.99      0.99    985600
+    accuracy                           0.99   5913600
+   macro avg       0.96      0.86      0.90   5913600
+weighted avg       0.99      0.99      0.99   5913600
 
 Confusion Matrix
-[[860699   1591]
- [  8652 114658]]
+[[5782308    7982]
+ [  34334   88976]]
