+++ .github/workflows/ci-cd-main.yml
@@ -30,7 +30,7 @@ jobs:
         uses: actions/checkout@v3
 
       - name: Check style
-      - uses: pre-commit/action@v2.0.0
+        uses: pre-commit/action@v2.0.0
   ci:
     needs: style
     if: github.event.pull_request.merged == true || github.event_name == 'push'
