### 2023-09-19 da3e793
chore: workflow for generating paper in new branch
