+++ .github/workflows/draft-pdf-rev.yml
@@ -17,11 +17,11 @@ jobs:
           journal: joss
           # This should be the path to the paper within your repo.
           paper-path: paper/paper.md
-      #- name: Upload
-      #  uses: actions/upload-artifact@v1
-      #  with:
-      #    name: paper
+      - name: Upload
+        uses: actions/upload-artifact@v1
+        with:
+          name: paper
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-      #    path: paper/paper.pdf
+          path: paper/paper-rev.pdf
