+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @article{moody2001impact,
   author={Moody, G.B. and Mark, R.G.},
-  journal={IEEE Engineering in Medicine and Biology Magazine}, 
-  title={The impact of the MIT-BIH Arrhythmia Database}, 
+  journal={IEEE Engineering in Medicine and Biology Magazine},
+  title={The impact of the MIT-BIH Arrhythmia Database},
   year={2001},
   volume={20},
   number={3},
@@ -38,7 +38,7 @@
       Johnson, A. and
       Pollard, T. and
       Gow, B. and
-      Moody, B.}, 
+      Moody, B.},
   year={2023},
   journal={PhysioNet},
   doi={10.13026/9njx-6322}
@@ -107,7 +107,7 @@
   url={https://keras.io}
 }
 
-@article{Makowski_NeuroKit2_A_Python_2021,
+@article{NeuroKit2_2021,
   author = {
       Makowski, Dominique and
       Pham, Tam and
@@ -125,4 +125,39 @@
   title = {{NeuroKit2: A Python toolbox for neurophysiological signal processing}},
   volume = {53},
   year = {2021}
+}
+
+@article{SleepECG_2023,
+  doi = {10.21105/joss.05411},
+  url = {https://doi.org/10.21105/joss.05411},
+  year = {2023},
+  publisher = {The Open Journal},
+  volume = {8},
+  number = {86},
+  pages = {5411},
+  author = {Clemens Brunner and Florian Hofer},
+  title = {SleepECG: a Python package for sleep staging based on heart rate},
+  journal = {Journal of Open Source Software}
+}
+
+@article{PanTompkins1985,
+  author={Pan, Jiapu and Tompkins, Willis J.},
+  journal={IEEE Transactions on Biomedical Engineering},
+  title={A Real-Time QRS Detection Algorithm},
+  year={1985},
+  volume={BME-32},
+  number={3},
+  pages={230-236},
+  doi={10.1109/TBME.1985.325532}
+}
+
+@article{Hamilton2002,
+  author={Hamilton, P.},
+  booktitle={Computers in Cardiology},
+  title={Open source ECG analysis},
+  year={2002},
+  volume={},
+  number={},
+  pages={101-104},
+  doi={10.1109/CIC.2002.1166717}
 }
