+++ examples/model/result.txt
@@ -1,12 +1,12 @@
               precision    recall  f1-score   support
 
-           0       0.99      1.00      1.00   5790290
-           1       0.92      0.72      0.81    123310
+           0       1.00      1.00      1.00   5803163
+           1       0.93      0.91      0.92    123637
 
-    accuracy                           0.99   5913600
-   macro avg       0.96      0.86      0.90   5913600
-weighted avg       0.99      0.99      0.99   5913600
+    accuracy                           1.00   5926800
+   macro avg       0.96      0.96      0.96   5926800
+weighted avg       1.00      1.00      1.00   5926800
 
 Confusion Matrix
-[[5782308    7982]
- [  34334   88976]]
+[[5794101    9062]
+ [  10690  112947]]
