+++ paper/paper.md
@@ -25,11 +25,7 @@ Electrocardiogram (ECG) signals represent the electrical activity of the heart a
 
 # Statement of need
 
-Recent advances in deep learning have made it possible to accomplish tasks previously unattainable through the implementation of conventional algorithms.
-
-The capabilities of conventional approaches for analyzing ECG signals are limited. Therefore, it is imperative to develop data-driven models capable of automatically identifying patterns from vast heterogeneous data.
-
-`Pyheartlib` supports this goal and, unlike popular packages such as NeuroKit2, does not implement pre-defined algorithms. It instead provides data processing modules to assist researchers with the implementation of their own models for various tasks. For example, SleepECG [@SleepECG_2023] and Neurokit2 [@NeuroKit2_2021] have implemented several conventional methods, such as the Pan–Tompkins [@PanTompkins1985] and Hamilton [@Hamilton2002] algorithms, for the R-peak detection task. However, the same task can be accomplished through deep learning. Various models can be trained by utilizing the readily available data processing modules in `Pyheartlib`. An example model is provided with this package that performs this task. `Pyheartlib` comprises several dedicated classes for different use cases.
+Recent advances in deep learning have made it possible to accomplish tasks previously unattainable through the implementation of conventional algorithms. The capabilities of conventional approaches for analyzing ECG signals are limited. Therefore, it is imperative to develop data-driven models capable of automatically identifying patterns from vast heterogeneous data. `Pyheartlib` supports this goal and, unlike popular packages such as NeuroKit2 [@NeuroKit2_2021], does not implement pre-defined algorithms. It instead provides data processing modules to assist researchers with the implementation of their own models for various tasks. For example, SleepECG [@SleepECG_2023] and Neurokit2 [@NeuroKit2_2021] have implemented several conventional methods, such as the Pan–Tompkins [@PanTompkins1985] and Hamilton [@Hamilton2002] algorithms, for the R-peak detection task. However, the same task can be accomplished through deep learning. Various models can be trained by utilizing the readily available data processing modules in `Pyheartlib`. An example model is provided with this package that performs this task. `Pyheartlib` comprises several dedicated classes for different use cases.
 
 For the heartbeat classification task, which typically requires segmented and annotated heartbeat waveforms, the software preprocesses the input data and delivers datasets comprising waveforms and features. Features can be computed from waveform and RR-intervals.
 
