+++ paper/paper.md
@@ -31,11 +31,11 @@ For the heartbeat classification task, which typically requires segmented and an
 
 For the classification of signal excerpts, e.g., arrhythmia classification, the software is designed to store metadata about the excerpts in order to reduce memory usage significantly, especially in situations that each excerpt is being used only once during model training. By adjusting the parameters appropriately, the package provides a substantial quantity of data samples for training deep learning models. Moreover, it is feasible to incorporate RR-intervals in addition to waveforms and their computed features.
 
-Another use case is when each excerpt has to be divided into smaller sub-segments each with a specific label, e.g., r-peak detection. The package delivers data samples by storing metadata about the excerpts and providing lists of labels as annotations for the excerpts.
+Another use case is when each excerpt has to be divided into smaller sub-segments each with a specific label, e.g., R-peak detection. The package delivers data samples by storing metadata about the excerpts and providing lists of labels as annotations for the excerpts.
 
 `Pyheartlib` is easy to use and its documentation contains examples for different use cases. \autoref{fig:example} illustrates an example outcome of a deep learning model that was trained with the help of `Pyheartlib` and Keras [@chollet2015keras] to detect R-peaks.
 
-![R-peak detection using deep learning. In this example use case, the Pyheartlib package was utilized for processing the signals for training and testing a deep learning model that was implemented using the Keras library. This figure illustrates the outcome of the model on the test dataset. Additional information about this example model is available on the project repository and its documentation. \label{fig:example}](mis.png){ width=95% }
+![R-peak detection using deep learning. In this example use case, the Pyheartlib package was utilized for processing the signals for training and testing a deep learning model that was implemented using the Keras library. This figure illustrates the outcome of the model on the test dataset. Additional information about this example model is available on the project repository and its documentation. \label{fig:example}](mis.png){ width=98% }
 
 This package utilizes the WFDB Python package to read data, therefore it supports the WFDB format. Recordings with different formats can be converted using the WFDB Python package [@moody2001impact;@goldberger2000physiobank;@wfdb-python;@wfdb-software-package].
 
