### 2023-09-21 a10d75a
chore: improve example model
