+++ examples/model/inference.py
@@ -11,6 +11,7 @@
 
 # This file can be used to do inference using the trained example model
 import os
+import textwrap
 
 import numpy as np
 import tensorflow as tf
@@ -36,7 +37,7 @@ annotated_records_test, samples_info_test = load_data(test_data)
 print(len(samples_info_test))
 from pyheartlib.data_rpeak import ECGSequence  # noqa: E402
 
-batch_size = 32
+batch_size = 128
 testseq = ECGSequence(
     annotated_records_test,
     samples_info_test,
@@ -59,10 +60,75 @@ assert labels_true.shape == labels_pred.shape
 
 # print classification_report and confusion_matrix
 true = labels_true.flatten()
-pre = labels_pred.flatten()
-print(classification_report(true, pre))
-print(confusion_matrix(true, pre))
+pred = labels_pred.flatten()
+print(classification_report(true, pred))
+print(confusion_matrix(true, pred))
 with open("./result.txt", "w") as f:
-    print(classification_report(true, pre), file=f)
+    title = (
+        "************************************************************"
+        "Without considering adjacency of ground truth and prediction"
+    )
+    print(textwrap.fill(title, width=60, replace_whitespace=False), file=f)
+    print("\n", file=f)
+    print(classification_report(true, pred), file=f)
     print("Confusion Matrix", file=f)
-    print(confusion_matrix(true, pre), file=f)
+    print(confusion_matrix(true, pred), file=f)
+    print("\n", file=f)
+
+
+def post_tolerance(true, pred):
+    # Give a tolerance for nearby segments (False positive)
+    # when a False Positve is very adjacent to ground truth peak
+    new_pred = np.copy(pred)
+    for indx, _ in enumerate(pred):
+        if pred[indx] == 1:
+            if true[indx] != 1:
+                if true[indx + 1] == 1:
+                    new_pred[indx + 1] = 1
+                    new_pred[indx] = 0
+                elif true[indx - 1] == 1:
+                    new_pred[indx - 1] = 1
+                    new_pred[indx] = 0
+                elif true[indx + 2] == 1:
+                    new_pred[indx + 2] = 1
+                    new_pred[indx] = 0
+                elif true[indx - 2] == 1:
+                    new_pred[indx - 2] = 1
+                    new_pred[indx] = 0
+                if true[indx + 3] == 1:
+                    new_pred[indx + 3] = 1
+                    new_pred[indx] = 0
+                elif true[indx - 3] == 1:
+                    new_pred[indx - 3] = 1
+                    new_pred[indx] = 0
+                if true[indx + 4] == 1:
+                    new_pred[indx + 4] = 1
+                    new_pred[indx] = 0
+                elif true[indx - 4] == 1:
+                    new_pred[indx - 4] = 1
+                    new_pred[indx] = 0
+                if true[indx + 5] == 1:
+                    new_pred[indx + 5] = 1
+                    new_pred[indx] = 0
+                elif true[indx - 5] == 1:
+                    new_pred[indx - 5] = 1
+                    new_pred[indx] = 0
+    return new_pred
+
+
+# print classification_report and confusion_matrix
+new_pred = post_tolerance(true, pred)
+pred = new_pred
+print(classification_report(true, pred))
+print(confusion_matrix(true, pred))
+with open("./result.txt", "a") as f:
+    title = (
+        "************************************************************"
+        "The results below were obtained by considering a very small"
+        "  tolerance for the adjacency of ground truth and prediction"
+    )
+    print(textwrap.fill(title, width=60, replace_whitespace=False), file=f)
+    print("\n", file=f)
+    print(classification_report(true, pred), file=f)
+    print("Confusion Matrix", file=f)
+    print(confusion_matrix(true, pred), file=f)
