### 2023-09-21 devnums/pyheartlib#8 @devnums
Thanks for your comment. The file has been updated. Please check it on the branch: paper
[contributing.md](https://github.com/devnums/pyheartlib/blob/paper/CONTRIBUTING.md)
