### 2023-09-21 devnums/pyheartlib#9 @devnums
Thanks for the comment. I have added more information to the figure. I have also added additional figures to the repository.
Please check the [new figure](https://github.com/devnums/pyheartlib/blob/paper-rev/examples/model/plots/mis.png).

The figures show the true peaks according to the manual annotations of the MIT-BIH dataset.

A false positive means that a sub-segment is predicted to contain an R-peak, but in reality it does not. The figures show that this is usually the case for nearby points close to the true R-peak. False positives are shown in orange color.
When false positives are close to the true R-peaks, we can say that the prediction is right. I have calculated the results considering this fact. [Results](https://github.com/devnums/pyheartlib/blob/paper-rev/examples/model/result.txt)
It is possible to improve detection with a post-processing algorithm in these situations. Sometimes, these false positives are caused by a small mistake in the manual annotations.

False negatives are when the model misses to detect a true R-peak.

Although this is just an example, the model predicts with high accuracy. It can be improved to achieve more accurate predictions and detect almost all of the R-peaks.
