### 2023-09-21 devnums/pyheartlib#10 @devnums
Thank you for your suggestion. I have edited the paper, please take a look at it [here](https://github.com/devnums/pyheartlib/blob/paper-rev/paper/paper.pdf) and let me know if that's what you mean.
