### 2023-10-03 devnums/pyheartlib#8 @LegrandNico
It looks like the discord link is broken, maybe this could be removed?
