### 2023-10-03 #5792 @LegrandNico
Thank you @devnums for your work on this. I have commented on some of the issues. Currently, the paper is still missing a side-by-side comparison between the deep learning methods that can be used and the most traditional peak detection algorithm that can be found in other packages. I think that is the main question that a potential user would have: why use a more complicated model if not for better performance/accuracy? But JOSS is not intended to host a method paper that would go too deep in such a comparison so I will let @britta-wstnr decide whether this is something that should be added to the paper or not.
