### 2023-10-05 b1b808b
chore: fix the discord link
