### 2023-10-05 devnums/pyheartlib#8 @devnums
You are right, I have created a new permanent link. [CONTRIBUTING.md](https://github.com/devnums/pyheartlib/blob/paper-rev/CONTRIBUTING.md)
