### 2023-10-19 #5792 @britta-wstnr
@LegrandNico and @devnums, I looked into it and it seems to me like there is a motivation in the paper on which gap this toolbox is filling as compared to other toolboxes (i.e., there is no other toolboxes allowing deep learning models to be utilized). Such a "statement of need" is necessary, but to show that a toolbox is performing better in comparison to other toolboxes is not. As you already suspected, @LegrandNico, this is beyond the scope of the JOSS paper (but can still be a good idea to show e.g. in the documentation or a separate paper as users indeed might ask this question).
So I guess the question then is: do you agree that there is a statement of need/statement of relation towards existing toolboxes in the paper? @LegrandNico
Also curious what the other reviewers will say @BSingstad and @cudmore!
