### 2023-10-23 devnums/pyheartlib#11 @cudmore
Provide jupyter notebooks with your example dataset and model code
In your readthedoc web documentation you have html files with pretty large code blocks and copy/paste to a users .py file is too cumbersome and error prone. Would be best to provide examples dataset and model as jupyter notebooks so users can git clone the repo and run them locally with `jupyter notebook`.

For one example
  https://pyheartlib.readthedocs.io/en/latest/examples/dataset/arrhythmia.html

Ah, now I see you have an `example/` folder in the repo with .py files with the source code shown in docs webpage.

Would be nice to provide this in one place, maybe as a jupyter notebook within the docs? Or at least mention the py files are available in the repo (from within the web docs).
