### 2023-10-23 devnums/pyheartlib#13 @cudmore
Can you provide more details on intermediate steps in your example model(s) pipeline?
Can we get more details on the intermediate steps in your analysis pipeline?

For example, your web docs have a model page
   https://pyheartlib.readthedocs.io/en/latest/examples/model/model.html

This page first links 'beat detection' and has headings :
 - Data Preparation
 - Training
 - Inference
 - Results

It results in a [plot](https://github.com/devnums/pyheartlib/blob/main/examples/model/plots/mis.png).

Can you provide more detailed documentation about each of these intermediate steps? I would treat each of your headings/steps like a function. For example: what is the input, what is the computation, and what is the output?

I suggest this documentation could be a jupyter notebook with descriptive markdown, code blocks, and importantly **intermediate plots** of the data starting with the raw data.

Basically, what exactly is happening at and between  each of the steps including 'data preparation', 'training', 'inference', 'results', and especially the final plot? What is the input, computation, and output of each of these steps?
