### 2023-10-23 devnums/pyheartlib#14 @cudmore
What is the relationship between examples dataset and model?
You describe how to create a number of [dataset](https://pyheartlib.readthedocs.io/en/latest/examples/dataset/dataset.html) including: arrhythmia, inter-patient beat, intra-patient beat, and Rpeak.

You then describe how to create a [beat detection](https://pyheartlib.readthedocs.io/en/latest/examples/model/model.html) model in a few steps like: data preparation, training, inference, and results.

As a user, I do not understand the connection between me running the code to create these 4 different **dataset(s)** and then the **model** code with an additional 4 steps.

Can you create a new page(s) in your documentation to explain this to me?

When I run the dataset code, is that used in the model code? If so, how?
