### 2023-10-23 #5792 @cudmore
Thank you for your excellent manuscript and important Python package.

### I have made a number of GitHug issues I would like addressed ...

   https://github.com/devnums/pyheartlib/issues/11

   https://github.com/devnums/pyheartlib/issues/12

   https://github.com/devnums/pyheartlib/issues/13

   https://github.com/devnums/pyheartlib/issues/14

### Here, I provide some critiques to improve your manuscript that are outside of github issues.

#### Topics mentioned in the manuscript that need to be added to the online documentation

As I read your manuscript, there are a number of details mentioned that I do not see in the online documentation. These details may be documented at the code level with docstring(s) but need to be broken out into user dcumentation.

I would like to see the following points expanded in the online documentation. If you incorporate these suggestions, I believe the Python package will be substantially more usefull.

1) You mention WFDB and NeuroKit python packages. You then potentially use them to load data. What does this raw data look like? Please provide some plots. Once your code solves the problem, what do the results look like? Some plots comparing the original data with your new results would be great.

2) "For the heartbeat classification task, which typically requires segmented and annotated heartbeat waveforms, the software preprocesses the input data and delivers datasets comprising waveforms and features"

What does the input for the 'heartbeat classification task' look like? Can we see the input data, segmented and heartbeat waveforms? Can we see "waveforms and features", etc, etc.? What is 'waveforms' and 'features'? Please provide us with more info on these objects/classes/data-strutures?

3) "For the classification of signal excerpts, e.g., arrhythmia classification, the software is designed to store metadata about the excerpts in order to reduce memory usage significantly, especially in situations that each excerpt is being used only once during model training"

Can we see an example of the "classification of signal excerpts"? What does a 'signal excerpt' look like? What does the classification of a signal excerpt look like? Can you also document how the software is "designed to store metadata"? What metadata are you storing about excerpts? Why do you need to reduce memory usage about an excerpt? Why is it a problem that each excerpt is being used only once during model training?

4) "By adjusting the parameters appropriately, the package provides a substantial quantity of data samples for training deep learning models. Moreover, it is feasible to incorporate RR-intervals in addition to waveforms and their computed features."

You provide a lot of example code but it is unclear to me where the 'parameters' are that a user can adjust 'apporpriately'? How does one incorporate 'RR-intervals', 'waveforms', and their 'computed features'. Can we get a jupyter notebook showing us how to do this?

5) "Another use case is when each excerpt has to be divided into smaller sub-segments each with a specific label, e.g., r-peak detection. The package delivers data samples by storing metadata about the excerpts and providing lists of labels as annotations for the excerpts."

I am following you here but you need to guide the reader with more detailed readthedoc documentation. Can we get an example where a user would 'divide each excerpt into smaller sub-segments - each with a specific label'? How does that relate to r-peak detection?

If most of my above comments are addressed, I think the final statement might be parsable, e.g. "The package delivers data samples by storing metadata about the excerpts and providing lists of labels as annotations for the excerpts."

**Summary.** Your manuscript, code, and documentation is excellent. I would like to see the above mentioned topics addressed in the online documentation. Basically, a number of more descriptive tutorials. Overall, very good work.
