### 2023-10-26 79b88fc
chore: replace example py files with notebooks
