+++ examples/model/inference-plot.py
@@ -1,16 +1,6 @@
-#############################################################################
-# Copyright (c) 2023 Pyheartlib team. - All Rights Reserved                 #
-# Project repo: https://github.com/devnums/pyheartlib                       #
-# Contact: devnums.code@gmail.com                                           #
-#                                                                           #
-# This file is part of the Pyheartlib project.                              #
-# To see the complete LICENSE file visit:                                   #
-# https://github.com/devnums/pyheartlib/blob/main/LICENSE                   #
-#############################################################################
-
-
-# This file can be used to do inference using the trained example model
+# Inference using the trained example model -- plots
 import os
+import pathlib
 
 import numpy as np
 import tensorflow as tf
@@ -22,7 +12,7 @@ cdir = os.path.dirname(os.path.abspath(__file__))
 os.chdir(cdir)
 print("Current directory changed to:\n", cdir)
 
-data_dir = "../../data"
+data_dir = "data"
 test_data = os.path.join(data_dir, "test.rpeak")
 mdl_checkpoint = os.path.join(os.getcwd(), "checkpoint/keras.exp")
 interval_value = 6
@@ -171,5 +161,6 @@ for i in tqdm(rang):  # range(round(len(samples_info_test)/batch_size)):
         plt.rcParams.update({"font.size": 12})
         plt.rcParams["axes.axisbelow"] = True
         # save fig
+        pathlib.Path("plots").mkdir(parents=True, exist_ok=True)
         fig_path = "./plots/" + str(i) + ".png"
         plt.savefig(fig_path, dpi=600)
