+++ docs/examples/model/RPEAK_DETECTION.rst
@@ -5,7 +5,12 @@ R-peak Detection
 This example contains the code for training a deep learning model for the R-peak detection task. As the task is not complex, it can be trained with high accuracy on a small amount of data.
 In this example, the electrocardiogram signals are segmented into ten seconds ecxerpts. Each excerpt corresponds to an annotation list of length 600 containing zeros and ones. A subsegment with an R-peak corresponds to one in the annotation list.
 
-The following steps prepare the data and train the model.
+This example is available on `GitHub <https://github.com/devnums/pyheartlib/blob/main/examples/>`_.
+
+.. raw:: html
+
+    <a target="_blank" href="https://colab.research.google.com/github/devnums/pyheartlib/blob/main/examples/model/rpeak_detection.ipynb">
+      <img src="https://colab.research.google.com/assets/colab-badge.svg" alt="Open In Colab"/></a>
 
 
 Data preparation
