+++ src/pyheartlib/beat_info.py
@@ -26,44 +26,45 @@ class BeatInfo:
     Parameters
     ----------
     beat_loc : int
-        Index of beat in the rpeaks locations list,
-        considering pre-RR and post-RR values.
+        Index of beat in the R-peaks locations list, considering preceding
+        R-peak locations and subsequent R-peak locations.
     fs : int, optional
         Sampling rate, by default 360
     in_ms : bool, optional
-        Whether to calculate rr-intervals in time(miliseconds) or samples,
+        Whether to calculate RR-intervals in time (miliseconds) or samples,
         by default True
 
     Attributes
     ----------
     beat_loc : int
-        Index of beat in the rpeaks locations list,
-        considering pre-RR and post-RR values.
+        Index of beat in the R-peaks locations list, considering preceding
+        R-peak locations and subsequent R-peak locations.
     fs : int, optional
         Sampling rate, by default 360
     in_ms : bool, optional
-        Whether to calculate rr-intervals in time(miliseconds) or samples,
+        Whether to calculate RR-intervals in time (miliseconds) or samples,
         by default True
     whole_waveform : list
-        Whole waveform.
+        Whole waveform, that is not trimmed.
     bwaveform : list
-        Trimmed beat waveform. Selected as a segment of
-        the whole provided waveform.
+        Segmented beat waveform. Selected as a segment of
+        the whole waveform.
     start_idx: int
-        The index of the first sample of the waveform according to
-        the original signal.
+        The index of the first sample (onset) of the waveform
+        on the original signal.
     rri : list
-        RR intervals in ms if in_ms=True.
+        List of RR-intervals, in miliseconds if the in_ms parameter
+        has been set to True.
     rri_smpl : list
-        RR intervals in samples.
+        RR-intervals in samples.
     sdrri : list
-        Successive RRI differences.
+        Successive RR interval (RRI) differences.
     avail_features : list
-        List of fetaure names (strings) that already have a
-        definition in the class.
+        List of features names (strings) that already have a
+        definition (predefined features).
     features : dict
-        Dictionary with keys(strings) equal to feature names and dict values
-        equal to features values.
+        Dictionary containing computed features. Keys are feature names
+        and values are features values.
     """
 
     def __init__(self, beat_loc=10, fs=360, in_ms=True):
@@ -109,12 +110,12 @@ class BeatInfo:
             # self.__plot_wf() #plot for debug
 
     def available_features(self):
-        """Returns availble features that can be computed.
+        """Returns available features that can be computed.
 
         Returns
         -------
         list
-            List containing features names as strings.
+            List containing features names.
         """
         avail_feats = [f for f in dir(self) if f.startswith("F_")]
         return avail_feats
@@ -125,7 +126,8 @@ class BeatInfo:
         Parameters
         ----------
         new_features : list
-            Names of new features. Such as: [new_feature_1, new_feature_2]
+            Names of new features in a list.
+            Such as: [new_feature_1, new_feature_2]
         """
         for new_feature in new_features:
             setattr(
@@ -133,12 +135,12 @@ class BeatInfo:
             )
 
     def select_features(self, features):
-        """Select features.
+        """Select desired features for computation.
 
         Parameters
         ----------
         features : list
-            List of feature names(strings).
+            List of features names (strings).
         """
         self.selected_features_names = features
 
[…]
