+++ src/pyheartlib/config.yaml
@@ -14,6 +14,7 @@
 # Required fields
 DATA_DIR: "mit-bih-arrhythmia-database-1.0.0/"
 SAMPLING_RATE: 360
+CHANNEL: "MLII"
 BEAT_TYPES: ['N', 'L', 'R', 'j', 'e', 'V', 'E', 'A', 'S', 'a', 'J', 'F', 'f', '/', 'Q']
 RHYTHM_TYPES: ['(AB', '(AFIB', '(AFL', '(B', '(BII', '(IVR', '(N', '(NOD', '(P', '(PREX', '(SBR', '(SVTA', '(T', '(VFL', '(VT']
 
