### 2023-11-08 aab01af
refactor: a few changes in names, improve docs
