+++ src/pyheartlib/data_rpeak.py
@@ -25,24 +25,56 @@ from pyheartlib.features import get_wf_feats  # noqa: E402
 
 
 class RpeakData(Data, DataSeq):
-    """Provides signal data with annotations as a list.
+    """Processes ECG records to make a dataset holding records along with
+    metadata about signal excerpts.
+
+    It has a method that can generate metadata for signal excerpts.
+    The metadata are generated using the sliding window approach.
+    For each excerpt of a signal the onset, offset, and its annotation
+    is recorded. The metadata list for an excerpt is structured as:
+    [record_id, onset, offset, annotation].
+    Annotation for an excerpt is a list of zeros, except for any interval
+    where there is an R-peak label.
+    Example metadata for an excerpt:
+    [10, 500, 800, [0, 0, 0, 'N', 0, 'N', 0, 'N', 0, ...]]
 
     Parameters
     ----------
     base_path : str, optional
-        Path of main directory for loading and saving data, by default None
+        Path of the main directory for storing the original and
+        processed data, by default None
     remove_bl : bool, optional
-        If True, the baseline is removed from the raw signals
-        before extracting beat excerpts, by default False
+        If True, the baseline wander is removed from the original signals
+        prior to extracting excerpts, by default False
     lowpass : bool, optional
-        Whether to apply low pass filtering to the raw signals,
+        Whether or not to apply low-pass filter to the original signals,
         by default False
     cutoff : int, optional
-        Parameter of the low pass filter, by default 45
+        Parameter of the low pass-filter, by default 45
     order : int, optional
-        Parameter of the low pass filter, by default 15
+        Parameter of the low pass-filter, by default 15
     progress_bar : bool, optional
-        If True, progress bar is shown, by default True
+        Whether to display a progress bar, by default True
+
+    Example
+    -------
+    >>> from pyheartlib.data_rpeak import RpeakData
+    >>> # Make an instance of the RpeakData
+    >>> rpeak_data = RpeakData(
+    >>>     base_path="data", remove_bl=False, lowpass=False,
+    >>>     progress_bar=False)
+    >>> # Define records
+    >>> train_set = [201, 203]
+    >>> # Create the dataset
+    >>> # The win_size specifies the length of the excerpts
+    >>> rpeak_data.save_samples(
+    >>>     rec_list=train_set,
+    >>>     file_name="train.rpeak",
+    >>>     win_size=5 * 360,
+    >>>     stride=360,
+    >>>     interval=72,
+    >>> )
+
     """
 
     def __init__(
@@ -64,25 +96,26 @@ class RpeakData(Data, DataSeq):
         DataSeq.progress_bar = not progress_bar
 
     def full_annotate(self, record):
-        """Fully annotate a signal.
+        """Returns a signal along with an annotation of the same length.
 
         Parameters
         ----------
         record : dict
-            Record as a dictionary with keys: 'signal', 'r_locations',
-            'r_labels', 'rhythms', 'rhythms_locations'.
+            Record as a dictionary with keys: *signal*, *r_locations*,
+            *r_labels*, *rhythms*, *rhythms_locations*.
 
         Returns
         -------
-        list
-            A list of signal and full_ann: [signal, full_ann].
+        tuple
+            Two items: (signal, full_ann).
 
             First element is the original signal (1D ndarray).
 
             Second element is a list that has the same length as
             the original signal with zero elements except at any
-            rpeak index which has the rpeak annotation instead.
-            [00000N00000000000000L00000...]
+            R-peak index which has the R-peak label instead.
+            E.g.: [0,0,0,'N',0,'N',0,'N', ...]
+
         """
 
         signal, r_locations, r_labels, _, _ = record.values()
@@ -90,37 +123,56 @@ class RpeakData(Data, DataSeq):
         for i, loc in enumerate(r_locations):
[…]
