+++ src/pyheartlib/data_arrhythmia.py
@@ -1,291 +1,25 @@
-#############################################################################
-# Copyright (c) 2023 Pyheartlib team. - All Rights Reserved                 #
-# Project repo: https://github.com/devnums/pyheartlib                       #
-# Contact: devnums.code@gmail.com                                           #
-#                                                                           #
-# This file is part of the Pyheartlib project.                              #
-# To see the complete LICENSE file visit:                                   #
-# https://github.com/devnums/pyheartlib/blob/main/LICENSE                   #
-#############################################################################
+from warnings import warn
+
+from pyheartlib.data_rhythm import RhythmData
+
+msg = (
+    "Module data_arrhythmia will be deprecated in future versions. "
+    "Use module data_rhythm instead."
+)
+warn(
+    msg,
+    DeprecationWarning,
+    stacklevel=2,
+)
+
+msg = "ArrhythmiaData class will be deprecated. Use RhythmData class instead."
 
 
-import os
-
-os.environ["TF_CPP_MIN_LOG_LEVEL"] = "3"
-import math  # noqa: E402
-
-import numpy as np  # noqa: E402
-from tensorflow.keras.utils import Sequence  # noqa: E402
-from tqdm import tqdm  # noqa: E402
-
-from pyheartlib.data import Data, DataSeq  # noqa: E402
-from pyheartlib.features import get_hrv_features, get_wf_feats  # noqa: E402
-
-# import time  # noqa: E402
-
-
-class ArrhythmiaData(Data, DataSeq):
-    """Provides signal data with single value annotations.
-
-    Parameters
-    ----------
-    base_path : str, optional
-        Path of main directory for loading and saving data, by default None
-    remove_bl : bool, optional
-        If True, baseline will be removed from the raw signals
-        before extracting beat excerpts, by default False
-    lowpass : bool, optional
-        Whether to apply low pass filtering to the raw signals,
-        by default False
-    cutoff : int, optional
-        Parameter of the low pass filter, by default 45
-    order : int, optional
-        Parameter of the low pass filter, by default 15
-    """
-
-    def __init__(
-        self,
-        base_path=None,
-        remove_bl=False,
-        lowpass=False,
-        cutoff=45,
-        order=15,
-        progress_bar=True,
-    ):
-        super().__init__(
-            base_path,
-            remove_bl,
-            lowpass,
-            cutoff,
-            order,
+class ArrhythmiaData(RhythmData):
+    def __init__(self, *args, **kwargs):
+        warn(
+            msg,
+            DeprecationWarning,
+            stacklevel=2,
         )
-        DataSeq.progress_bar = not progress_bar
-
-    def full_annotate(self, record):
-        """Fully annotate a signal.
-
-        Parameters
-        ----------
-        record : dict
-            Record as a dictionary with keys: 'signal', 'r_locations',
-            'r_labels', 'rhythms', 'rhythms_locations'.
-
-        Returns
-        -------
-        list
-            A list of signal and full_ann: [signal, full_ann].
-
-            First element is the original signal (1D ndarray).
-
-            Second element is a list that has the same length as
-            the original signal with arrhythmia types at each index:
-            ['(N','(N', '(N','AFIB','AFIB','AFIB',...].
-        """
-
-        signal, _, _, rhythms, rhythms_locations = record.values()
-        sig_length = len(signal)
-        full_ann = []
-        full_ann = ["unlab"] * len(signal)
-        for i in range(len(rhythms_locations)):
-            remained = sig_length - rhythms_locations[i]
-            full_ann[rhythms_locations[i] :] = [rhythms[i]] * remained
-        record_full = [signal, full_ann]
-        return record_full
-
-    def make_samples_info(
[…]
