+++ src/pyheartlib/data_rhythm.py
@@ -1,0 +1,339 @@
+#############################################################################
+# Copyright (c) 2023 Pyheartlib team. - All Rights Reserved                 #
+# Project repo: https://github.com/devnums/pyheartlib                       #
+# Contact: devnums.code@gmail.com                                           #
+#                                                                           #
+# This file is part of the Pyheartlib project.                              #
+# To see the complete LICENSE file visit:                                   #
+# https://github.com/devnums/pyheartlib/blob/main/LICENSE                   #
+#############################################################################
+
+
+import os
+
+os.environ["TF_CPP_MIN_LOG_LEVEL"] = "3"
+import math  # noqa: E402
+
+import numpy as np  # noqa: E402
+from tensorflow.keras.utils import Sequence  # noqa: E402
+from tqdm import tqdm  # noqa: E402
+
+from pyheartlib.data import Data, DataSeq  # noqa: E402
+from pyheartlib.features import get_hrv_features, get_wf_feats  # noqa: E402
+
+# import time  # noqa: E402
+
+
+class RhythmData(Data, DataSeq):
+    """Processes ECG records to make a dataset holding records along with
+    metadata about signal excerpts.
+
+    It has a method that can generate metadata for signal excerpts.
+    The metadata are generated using the sliding window approach.
+    For each excerpt of a signal the onset, offset, and its annotation
+    is recorded. The metadata list for an excerpt is structured as:
+    [record_id, onset, offset, annotation].
+    Annotation for an excerpt is a single label.
+    Example metadata for an excerpt: [10, 500, 800, 'AFIB'].
+
+    Parameters
+    ----------
+    base_path : str, optional
+        Path of the main directory for storing the original and
+        processed data, by default None
+    remove_bl : bool, optional
+        If True, the baseline wander is removed from the original signals
+        prior to extracting excerpts, by default False
+    lowpass : bool, optional
+        Whether or not to apply low-pass filter to the original signals,
+        by default False
+    cutoff : int, optional
+        Parameter of the low pass-filter, by default 45
+    order : int, optional
+        Parameter of the low pass-filter, by default 15
+    progress_bar : bool, optional
+        Whether to display a progress bar, by default True
+
+    Example
+    -------
+    >>> from pyheartlib.data_rhythm import RhythmData
+    >>> # Make an instance of the RhythmData
+    >>> rhythm_data = RhythmData(
+    >>>     base_path="data", remove_bl=False, lowpass=False,
+    >>>     progress_bar=False)
+    >>> # Define records
+    >>> train_set = [201, 203]
+    >>> # Create the dataset
+    >>> rhythm_data.save_samples(
+    >>>   rec_list=train_set, file_name="train.arr", win_size=3600, stride=64
+    >>> )
+
+    """
+
+    def __init__(
+        self,
+        base_path=None,
+        remove_bl=False,
+        lowpass=False,
+        cutoff=45,
+        order=15,
+        progress_bar=True,
+    ):
+        super().__init__(
+            base_path,
+            remove_bl,
+            lowpass,
+            cutoff,
+            order,
+        )
+        DataSeq.progress_bar = not progress_bar
+
+    def full_annotate(self, record):
+        """Returns a signal along with an annotation of the same length.
+
+        Parameters
+        ----------
+        record : dict
+            Record as a dictionary with keys: *signal*, *r_locations*,
+            *r_labels*, *rhythms*, *rhythms_locations*.
+
+        Returns
+        -------
+        tuple
+            Two items: (signal, full_ann).
+
+            First element is the original signal (1D ndarray).
+
+            Second element is a list that has the same length as
[…]
