### 2023-11-08 8096d61
chore: improve examples
