+++ ACKNOWLEDGEMENTS.txt
@@ -5,7 +5,7 @@ This package has utilized the following third party components.
 
 
 
-Start of 'THIRDPARTYLICENSES' generated by Python third_party_license_generator at 2023-09-13 17:40:19.499499
+Start of 'THIRDPARTYLICENSES' generated by Python third_party_license_generator at 2023-11-08 23:43:17.643542
 
 ----------------------------------------
 
@@ -5173,7 +5173,7 @@ POSSIBILITY OF SUCH DAMAGE.
 
 ----------------------------------------
 
-End of 'THIRDPARTYLICENSES' generated by Python third_party_license_generator at 2023-09-13 17:40:19.500758
+End of 'THIRDPARTYLICENSES' generated by Python third_party_license_generator at 2023-11-08 23:43:17.644153
 ----------------------------------------
 
 Package: cookiecutter
