+++ docs/quickstart.md
@@ -1,2 +1,0 @@
-```{include} ../QUICKSTART.md
-```
