+++ docs/examples/examples.rst
@@ -1,8 +1,16 @@
+========
 Examples
 ========
 
-.. toctree::
-    :glob:
+.. include:: examples_content.md
+    :parser: myst_parser.sphinx_
 
-    dataset/dataset.rst
-    model/model.rst
+.. toctree::
+    :titlesonly:
+    :hidden:
+    :maxdepth: 1
+
+    heartbeat
+    arrhythmia
+    rpeak
+    rpeak_detection
