+++ docs/examples/dataset/dataset.rst
@@ -1,8 +1,0 @@
-Dataset
-=======
-
-.. toctree::
-    :titlesonly:
-    :glob:
-
-    *
