+++ docs/examples/model/model.rst
@@ -1,7 +1,0 @@
-Model
-=====
-
-.. toctree::
-    :glob:
-
-    *

[…]
