### 2023-11-09 e71c77f
fix: rename same_samples method
