+++ src/pyheartlib/data_rpeak.py
@@ -20,6 +20,7 @@ from tqdm import tqdm  # noqa: E402
 
 from pyheartlib.data import Data, DataSeq  # noqa: E402
 from pyheartlib.features import get_wf_feats  # noqa: E402
+from pyheartlib.io import load_data  # noqa: E402
 
 # import time  # noqa: E402
 
@@ -67,7 +68,7 @@ class RpeakData(Data, DataSeq):
     >>> train_set = [201, 203]
     >>> # Create the dataset
     >>> # The win_size specifies the length of the excerpts
-    >>> rpeak_data.save_samples(
+    >>> rpeak_data.save_dataset(
     >>>     rec_list=train_set,
     >>>     file_name="train.rpeak",
     >>>     win_size=5 * 360,
@@ -372,3 +373,15 @@ class ECGSequence(Sequence):
 
     def compute_wf_feats(self, seq):
         return get_wf_feats(seq, self.interval)
+
+
+def load_dataset(file_path=None):
+    """Loads the dataset.
+
+    Parameters
+    ----------
+    file_path : str, optional
+        Path of the dataset, by default None
+
+    """
+    return load_data(file_path)
