+++ tests/test_rpeak.py
@@ -13,8 +13,7 @@ import numpy
 import pytest
 from dummy import DummyData
 
-from pyheartlib.data_rpeak import ECGSequence, RpeakData
-from pyheartlib.io import load_data
+from pyheartlib.data_rpeak import ECGSequence, RpeakData, load_dataset
 
 
 @pytest.fixture
@@ -89,14 +88,16 @@ def rpeakdata():
 
 
 def test_save_samples(rpeakdata):
-    ann, sam = rpeakdata.save_samples(
+    ann, sam = rpeakdata.save_dataset(
         rec_list=["dummy101", "dummy102"],
         file_name="tmp.rpeak",
         win_size=400,
         stride=200,
         return_ds=True,
     )
-    annotated_records, samples_info = load_data(test_data_dir + "/tmp.rpeak")
+    annotated_records, samples_info = load_dataset(
+        test_data_dir + "/tmp.rpeak"
+    )
     assert annotated_records[0]["r_labels"] == ann[0]["r_labels"]
     assert samples_info[0][2] == sam[0][2]
 
@@ -104,7 +105,7 @@ def test_save_samples(rpeakdata):
 @pytest.fixture
 def seq_generator1():
     obj = RpeakData(base_path=test_data_dir, remove_bl=False, lowpass=False)
-    ann, sam = obj.save_samples(
+    ann, sam = obj.save_dataset(
         rec_list=["dummy101", "dummy102"],
         file_name="tmp.rpeak",
         win_size=400,
@@ -145,7 +146,7 @@ def test_getitem1(seq_generator1):
 @pytest.fixture
 def seq_generator2():
     obj = RpeakData(base_path=test_data_dir, remove_bl=False, lowpass=False)
-    ann, sam = obj.save_samples(
+    ann, sam = obj.save_dataset(
         rec_list=["dummy101", "dummy102"],
         file_name="tmp.rpeak",
         win_size=400,
@@ -186,7 +187,7 @@ def test_getitem2(seq_generator2):
 @pytest.fixture
 def seq_generator3():
     obj = RpeakData(base_path=test_data_dir, remove_bl=False, lowpass=False)
-    ann, sam = obj.save_samples(
+    ann, sam = obj.save_dataset(
         rec_list=["dummy101", "dummy102"],
         file_name="tmp.rpeak",
         win_size=400,
