+++ docs/examples/result.txt
@@ -1,0 +1,35 @@
+************************************************************
+Without considering adjacency of ground truth and prediction
+
+
+              precision    recall  f1-score   support
+
+           0       1.00      1.00      1.00   5790290
+           1       0.92      0.92      0.92    123310
+
+    accuracy                           1.00   5913600
+   macro avg       0.96      0.96      0.96   5913600
+weighted avg       1.00      1.00      1.00   5913600
+
+Confusion Matrix
+[[5780883    9407]
+ [  10219  113091]]
+
+
+************************************************************
+The results below were obtained by considering a very small
+tolerance for the adjacency of ground truth and prediction
+
+
+              precision    recall  f1-score   support
+
+           0       1.00      1.00      1.00   5790290
+           1       1.00      0.98      0.99    123310
+
+    accuracy                           1.00   5913600
+   macro avg       1.00      0.99      0.99   5913600
+weighted avg       1.00      1.00      1.00   5913600
+
+Confusion Matrix
+[[5790060     230]
+ [   2460  120850]]
