+++ .github/workflows/ci-dev.yml
@@ -17,35 +17,39 @@ on:
 
 jobs:
   ci:
-
-    runs-on: ubuntu-latest
     strategy:
-      fail-fast: false
+      fail-fast: true
       matrix:
+        os: [ubuntu-20.04, ubuntu-22.04, ubuntu-latest, macos-latest, windows-latest]
         python-version: ["3.10", "3.11"]
 
+    runs-on: ${{ matrix.os }}
     steps:
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v3
         with:
           python-version: ${{ matrix.python-version }}
 
-      - name: Check-out repository
-        uses: actions/checkout@v3
+      - name: Echo
+        run: |
+            echo ***********************************
+            echo OS= ${{ matrix.os }}
+            echo Python= ${{ matrix.python-version }}
+            echo ***********************************
 
-      - name: Install poetry
-        uses: snok/install-poetry@v1
+      # - name: Check-out repository
+      #   uses: actions/checkout@v3
 
-      - name: Install package
-        run: poetry install
+      # - name: Install poetry
+      #   uses: snok/install-poetry@v1
 
-      - name: Test with pytest
-        run: poetry run pytest tests/ --cov=pyheartlib --cov-report=xml
+      # - name: Install package
+      #   run: poetry install
 
-      - name: Use Codecov to track coverage
-        uses: codecov/codecov-action@v3
-        with:
-          files: ./coverage.xml   # coverage report
+      # - name: Test with pytest
+      #   run: poetry run pytest tests/ --cov=pyheartlib --cov-report=xml
 
-      - name: Build documentation
-        run: poetry run make html --directory docs/
+      # - name: Use Codecov to track coverage
+      #   uses: codecov/codecov-action@v3
+      #   with:
+      #     files: ./coverage.xml   # coverage report
