+++ .github/workflows/ci-dev.yml
@@ -21,7 +21,7 @@ jobs:
       fail-fast: true
       matrix:
         os: [ubuntu-20.04, ubuntu-22.04, ubuntu-latest, macos-latest, windows-latest]
-        python-version: ["3.10", "3.11"]
+        python-version: ["3.10", "3.11", "3.12"]
 
     runs-on: ${{ matrix.os }}
     steps:
@@ -31,23 +31,19 @@ jobs:
           python-version: ${{ matrix.python-version }}
 
       - name: Echo
-        run: |
-            echo ***********************************
-            echo OS= ${{ matrix.os }}
-            echo Python= ${{ matrix.python-version }}
-            echo ***********************************
+        run: echo OS= ${{ matrix.os }} Python= ${{ matrix.python-version }}
 
-      # - name: Check-out repository
-      #   uses: actions/checkout@v3
+      - name: Check-out repository
+        uses: actions/checkout@v3
 
-      # - name: Install poetry
-      #   uses: snok/install-poetry@v1
+      - name: Install poetry
+        uses: snok/install-poetry@v1
 
-      # - name: Install package
-      #   run: poetry install
+      - name: Install package
+        run: poetry install
 
-      # - name: Test with pytest
-      #   run: poetry run pytest tests/ --cov=pyheartlib --cov-report=xml
+      - name: Test with pytest
+        run: poetry run pytest tests/ --cov=pyheartlib --cov-report=xml
 
       # - name: Use Codecov to track coverage
       #   uses: codecov/codecov-action@v3
