### 2023-11-09 d0bc777
build: edit python version requirement
