### 2023-11-09 d5d7668
ci: fail fast set to false
