+++ .github/workflows/ci-dev.yml
@@ -18,7 +18,7 @@ on:
 jobs:
   ci:
     strategy:
-      fail-fast: true
+      fail-fast: false
       matrix:
         os: [ubuntu-20.04, ubuntu-22.04, ubuntu-latest, macos-latest, windows-latest]
         python-version: ["3.10", "3.11", "3.12"]
