### 2023-11-09 0848090
buidl: revert python version reqs
