+++ .github/workflows/ci-dev.yml
@@ -20,7 +20,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-20.04, ubuntu-22.04, macos-latest, windows-latest]
+        os: [ubuntu-20.04, ubuntu-22.04, ubuntu-latest, macos-latest, windows-latest]
         python-version: ["3.10", "3.11"]
 
     runs-on: ${{ matrix.os }}
@@ -39,13 +39,19 @@ jobs:
       - name: Install poetry
         uses: snok/install-poetry@v1
 
-      - name: Install & Test ubuntu
+      - name: Install & Test Linux
         if: runner.os == 'Linux'
         run: |
             poetry install
             poetry run pytest tests/ --cov=pyheartlib --cov-report=xml
 
-      - name: Install & Test windows
+      - name: Install & Test macOS
+        if: runner.os == 'macOS'
+        run: |
+            poetry install
+            poetry run pytest tests/ --cov=pyheartlib --cov-report=xml
+
+      - name: Install & Test Windows
         if: runner.os == 'Windows'
         run: |
            C:\Users\runneradmin\.local\bin\poetry install
