+++ .github/workflows/ci-dev.yml
@@ -20,7 +20,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-20.04, ubuntu-22.04, macos-12, macos-13, macos-14, windows-latest]
+        os: [ubuntu-20.04, ubuntu-22.04, macos-12, macos-13, macos-14]
         python-version: ["3.10", "3.11"]
 
     runs-on: ${{ matrix.os }}
