+++ .github/workflows/ci-cd-main.yml
@@ -84,15 +84,6 @@ jobs:
       - name: Install package
         run: poetry install
 
-      - name: Use Python Semantic Release to prepare release
-        env:
-          # This token is created automatically by GH Actions
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: |
-            git config user.name github-actions
-            git config user.email github-actions@github.com
-            poetry run semantic-release publish
-
       - name: Test with pytest
         run: poetry run pytest tests/ --cov=pyheartlib --cov-report=xml
 
@@ -100,6 +91,24 @@ jobs:
         uses: codecov/codecov-action@v3
         with:
           files: ./coverage.xml   # coverage report
+
+      # This action uses Python Semantic Release v8
+      - name: Python Semantic Release
+        id: release
+        uses: python-semantic-release/python-semantic-release@v8.0.0
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+
+
+
+      # - name: Use Python Semantic Release to prepare release
+      #   env:
+      #     # This token is created automatically by GH Actions
+      #     GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      #   run: |
+      #       git config user.name github-actions
+      #       git config user.email github-actions@github.com
+      #       poetry run semantic-release publish
 
       - name: Publish to TestPyPI
         uses: pypa/gh-action-pypi-publish@release/v1
@@ -115,8 +124,8 @@ jobs:
             --extra-index-url https://pypi.org/simple \
             pyheartlib
 
-      - name: Publish to PyPI
-        uses: pypa/gh-action-pypi-publish@release/v1
-        with:
-          user: __token__
-          password: ${{ secrets.PYPI_API_TOKEN }}
+      # - name: Publish to PyPI
+      #   uses: pypa/gh-action-pypi-publish@release/v1
+      #   with:
+      #     user: __token__
+      #     password: ${{ secrets.PYPI_API_TOKEN }}
