### 2023-11-10 d31cab5
1.0.2
Automatically generated by python-semantic-release
