+++ CHANGELOG.md
@@ -1,128 +1,822 @@
-# Changelog
+# CHANGELOG
 
-<!--next-version-placeholder-->
+
+
+## v1.0.2 (2023-11-10)
+
+### Build
+
+* build: update file ([`b523c81`](https://github.com/devnums/pyheartlib/commit/b523c8141bf8de2e2d4361e5cfa25610cec1f98b))
+
+* build: edit python version requirement ([`d0bc777`](https://github.com/devnums/pyheartlib/commit/d0bc777d59f533b3a1507484728e8779d6bfbd06))
+
+* build: update deps ([`6751025`](https://github.com/devnums/pyheartlib/commit/675102532e8a92e55571ae997183f9a580e294bf))
+
+* build: update deps ([`136f3e4`](https://github.com/devnums/pyheartlib/commit/136f3e488f545bdedfaa36106614c05080612d61))
+
+* build: delete ([`8f4fb5a`](https://github.com/devnums/pyheartlib/commit/8f4fb5af214cae3c142094cb09a7d06ea43f0247))
+
+* build: update deps ([`e195f5f`](https://github.com/devnums/pyheartlib/commit/e195f5fc76e6766590361ea3614ccfcb64df6a07))
+
+* build: add new package ([`80b98c3`](https://github.com/devnums/pyheartlib/commit/80b98c377ed62ac9395a1872b3fbefd0d6763197))
+
+* build: increase file size upload ([`60825d3`](https://github.com/devnums/pyheartlib/commit/60825d319c785be7ee41b41f90c59888d6f2118b))
+
+* build: add new dev deps ([`8fafb5e`](https://github.com/devnums/pyheartlib/commit/8fafb5e2696897db5054dfba797e147384373e91))
+
+* build: delete and add lock file ([`a69e315`](https://github.com/devnums/pyheartlib/commit/a69e31563472dfc1c3fcc9dbc13467c2b0b143f0))
+
+### Chore
+
+* chore: rerun example notebooks ([`cda73e2`](https://github.com/devnums/pyheartlib/commit/cda73e22a9701176a2a847c56a032dcb448c215e))
+
+* chore:  fix command ([`ae38d57`](https://github.com/devnums/pyheartlib/commit/ae38d57d4e7ae33316a653bc3fbc1af774075509))
+
+* chore: edit file ([`501acf6`](https://github.com/devnums/pyheartlib/commit/501acf6c20ba8128e03a78c091a9b7134e4f3ba6))
+
+* chore: update file ([`26f18bf`](https://github.com/devnums/pyheartlib/commit/26f18bfcf1eb403304b3fd76c9e10f302fd87a18))
+
+* chore: update file ([`5f4ed33`](https://github.com/devnums/pyheartlib/commit/5f4ed3369c922d8f9171e1ce877670e8285ca8f6))
+
+* chore: add file ([`c00987a`](https://github.com/devnums/pyheartlib/commit/c00987abed96be04ac46b1b2799a0068715f971d))
+
+* chore: update examples ([`df21d95`](https://github.com/devnums/pyheartlib/commit/df21d95c5547f44c01c31d7d9d0915b4d42f26c6))
+
+* chore: edit and rename ([`a65e179`](https://github.com/devnums/pyheartlib/commit/a65e179e01ed577022a9bad7478c116ebbd1a1ac))
+
+* chore: delete file ([`b84c407`](https://github.com/devnums/pyheartlib/commit/b84c407532a5de6c4a1d7802ce8864562104433b))
+
[…]
