+++ .github/workflows/ci-cd-main.yml
@@ -84,31 +84,14 @@ jobs:
       - name: Install package
         run: poetry install
 
-      - name: Test with pytest
-        run: poetry run pytest tests/ --cov=pyheartlib --cov-report=xml
-
-      - name: Use Codecov to track coverage
-        uses: codecov/codecov-action@v3
-        with:
-          files: ./coverage.xml   # coverage report
-
-      # This action uses Python Semantic Release v8
-      - name: Python Semantic Release
-        id: release
-        uses: python-semantic-release/python-semantic-release@v8.0.0
-        with:
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-
-
-
-      # - name: Use Python Semantic Release to prepare release
-      #   env:
-      #     # This token is created automatically by GH Actions
-      #     GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-      #   run: |
-      #       git config user.name github-actions
-      #       git config user.email github-actions@github.com
-      #       poetry run semantic-release publish
+      - name: Use Python Semantic Release to prepare release
+        env:
+          # This token is created automatically by GH Actions
+          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+            git config user.name github-actions
+            git config user.email github-actions@github.com
+            poetry run semantic-release version
 
       - name: Publish to TestPyPI
         uses: pypa/gh-action-pypi-publish@release/v1
