+++ pyproject.toml
@@ -39,15 +39,15 @@ ruff = ">=0.0.290"
 sphinx-togglebutton = ">=0.3.2"
 
 [tool.semantic_release]
-version_toml = ["pyproject.toml:tool.poetry.version"]
-branch = "main"                                     # branch to make releases of
-changelog_file = "CHANGELOG.md"                     # changelog file
-build_command = "pip install poetry && poetry build"# build dists
-dist_path = "dist/"                                 # where to put dists
-upload_to_release = true                            # auto-create GitHub release
-upload_to_pypi = false                              # don't auto-upload to PyPI
-remove_dist = false                                 # don't remove dists
-patch_without_tag = true                            # patch release by default
+version_variable = "pyproject.toml:version" # version location
+branch = "main"                             # branch to make releases of
+changelog_file = "CHANGELOG.md"             # changelog file
+build_command = "poetry build"              # build dists
+dist_path = "dist/"                         # where to put dists
+upload_to_release = true                    # auto-create GitHub release
+upload_to_pypi = false                      # don't auto-upload to PyPI
+remove_dist = false                         # don't remove dists
+patch_without_tag = true                    # patch release by default
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
