+++ .github/workflows/ci-cd-main.yml
@@ -58,8 +58,8 @@ jobs:
       - name: Test with pytest
         run: poetry run pytest tests/ --cov=pyheartlib
 
-      - name: Build documentation
-        run: poetry run make html --directory docs/
+      # - name: Build documentation
+      #   run: poetry run make html --directory docs/
 
   cd:
     # Only run this job if the "ci" job passes
