+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 
 
+## v1.0.3 (2023-11-10)
+
+### Build
+
+* build: update ([`b17d089`](https://github.com/devnums/pyheartlib/commit/b17d089702c6d466486ffe12a1d398fc8cb0f83d))
+
+### Ci
+
+* ci: update ([`d2e8774`](https://github.com/devnums/pyheartlib/commit/d2e8774d79dc0d56720429d7cac221bec3040111))
+
+### Fix
+
+* fix: update ([`959b12d`](https://github.com/devnums/pyheartlib/commit/959b12d9975eccea721b701ed2036d896bb4ccd2))
+
+### Unknown
+
+* Merge branch &#39;main&#39; of https://github.com/devnums/pyheartlib ([`adbc4a8`](https://github.com/devnums/pyheartlib/commit/adbc4a8f19695c073947a11ba59b11a619441cc3))
+
+
 ## v1.0.2 (2023-11-10)
 
 ### Build
