### 2023-11-10 611123a
1.0.4
Automatically generated by python-semantic-release
