+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pyheartlib"
-version = "1.0.3"
+version = "1.0.4"
 description = "A Python package for processing electrocardiogram signals"
 authors = ["devnums <devnums.code@gmail.com>"]
 license = "AGPL-3.0-only"
@@ -40,7 +40,7 @@ sphinx-togglebutton = ">=0.3.2"
 
 [tool.semantic_release]
 #version_variable = "pyproject.toml:tool.poetry.version"
-version_toml = "pyproject.toml:tool.poetry.version"
+version_toml = ["pyproject.toml:tool.poetry.version"]
 branch = "main"
 changelog_file = "CHANGELOG.md"
 build_command = "poetry build"
@@ -53,5 +53,5 @@ commit_version_number = true
 version_source = "tag"
 
 [build-system]
-requires = ["poetry-core>=1.0.0"]
+requires = ["poetry-core>=1.6.0"]
 build-backend = "poetry.core.masonry.api"
