### 2023-11-10 60dddb9
1.0.5
Automatically generated by python-semantic-release
