### 2023-11-10 6ae7c8f
build: add a new dep
