+++ pyproject.toml
@@ -37,6 +37,7 @@ pre-commit = ">=3.4.0"
 isort = ">=5.12.0"
 ruff = ">=0.0.290"
 sphinx-togglebutton = ">=0.3.2"
+tomli = ">=2.0.1"
 
 [tool.semantic_release]
 #version_variable = "pyproject.toml:tool.poetry.version"
