+++ docs/requirements.txt
@@ -1,3 +1,4 @@
 myst-nb
 sphinx-autoapi
 sphinx-rtd-theme
+tomli
