### 2023-11-10 f77d4a1
1.0.6
Automatically generated by python-semantic-release
