### 2023-11-10 f810572
docs: add new dep to the requirements
