+++ docs/requirements.txt
@@ -1,4 +1,5 @@
 myst-nb
 sphinx-autoapi
 sphinx-rtd-theme
+sphinx_togglebutton
 tomli
