### 2023-11-11 f41c5b2
build: update version source
