+++ .github/workflows/ci-cd-main.yml
@@ -98,8 +98,6 @@ jobs:
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
 
-
-
       - name: Test with pytest
         run: poetry run pytest tests/ --cov=pyheartlib --cov-report=xml
 
