+++ ACKNOWLEDGEMENTS.txt
@@ -5,7 +5,7 @@ This package has utilized the following third party components.
 
 
 
-Start of 'THIRDPARTYLICENSES' generated by Python third_party_license_generator at 2023-11-08 23:43:17.643542
+Start of 'THIRDPARTYLICENSES' generated by Python third_party_license_generator at 2023-11-11 00:34:26.242045
 
 ----------------------------------------
 
@@ -441,7 +441,7 @@ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND 
 Package: contourpy
 License: BSD-3-clause
 Requires: numpy
-Author: Ian Thomas <ianthomas23@gmail.com>
+Author: None
 Home page: None
 
 BSD 3-Clause License
@@ -479,8 +479,8 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 Package: cycler
 License: BSD-3-clause
 Requires: n/a
-Author: Thomas A Caswell <matplotlib-users@python.org>
-Home page: https://github.com/matplotlib/cycler
+Author: (unknown) <Thomas A Caswell <matplotlib-users@python.org>>
+Home page: None
 
 Copyright (c) 2015, matplotlib project
 All rights reserved.
@@ -676,7 +676,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 Package: google-auth
 License: Apache-2.0
-Requires: cachetools, pyasn1-modules, six
+Requires: cachetools, pyasn1-modules, rsa
 Author: Google Cloud Platform <googleapis-packages@google.com>
 Home page: https://github.com/googleapis/google-auth-library-python
 
@@ -2002,8 +2002,8 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 Package: joblib
 License: BSD-3-clause
 Requires: n/a
-Author: Gael Varoquaux <gael.varoquaux@normalesup.org>
-Home page: https://joblib.readthedocs.io
+Author: (unknown) <Gael Varoquaux <gael.varoquaux@normalesup.org>>
+Home page: None
 
 BSD 3-Clause License
 
@@ -2043,193 +2043,79 @@ Requires: n/a
 Author: Keras team <keras-users@googlegroups.com>
 Home page: https://keras.io/
 
+NOTE: This module was missing a license file (despite listing a license name) so one has been auto-generated
+
 Apache License
-                           Version 2.0, January 2004
-                        http://www.apache.org/licenses/
+Version 2.0, January 2004
+http://www.apache.org/licenses/
 
-   TERMS AND CONDITIONS FOR USE, REPRODUCTION, AND DISTRIBUTION
+TERMS AND CONDITIONS FOR USE, REPRODUCTION, AND DISTRIBUTION
 
-   1. Definitions.
+1. Definitions.
 
-      "License" shall mean the terms and conditions for use, reproduction,
-      and distribution as defined by Sections 1 through 9 of this document.
+"License" shall mean the terms and conditions for use, reproduction, and distribution as defined by Sections 1 through 9 of this document.
 
-      "Licensor" shall mean the copyright owner or entity authorized by
-      the copyright owner that is granting the License.
+"Licensor" shall mean the copyright owner or entity authorized by the copyright owner that is granting the License.
 
-      "Legal Entity" shall mean the union of the acting entity and all
-      other entities that control, are controlled by, or are under common
-      control with that entity. For the purposes of this definition,
-      "control" means (i) the power, direct or indirect, to cause the
[…]
