+++ CHANGELOG.md
@@ -1,67 +1,5 @@
-# CHANGELOG
+# Changelog
 
-
-
-## v1.0.6 (2023-11-10)
-
-### Build
-
-* build: update field ([`336c3da`](https://github.com/devnums/pyheartlib/commit/336c3da44fa74a9fb58b99c880d1b1c7a0fbbf2e))
-
-* build: add a new dep ([`6ae7c8f`](https://github.com/devnums/pyheartlib/commit/6ae7c8fe597a85fa4e124044b6a88c5201064fff))
-
-### Ci
-
-* ci: update workflow ([`76d2814`](https://github.com/devnums/pyheartlib/commit/76d28143c516b55b10994619a8ac8daa0dfeaee3))
-
-### Documentation
-
-* docs: add requirement ([`43b86ff`](https://github.com/devnums/pyheartlib/commit/43b86ffbc0589144d5f3793fb0e313ed8b08d936))
-
-* docs: fix ([`a821e6f`](https://github.com/devnums/pyheartlib/commit/a821e6f140df57ed0ebe569454d81189d933cbf5))
-
-### Fix
-
-* fix: fix bug ([`1ea3f4e`](https://github.com/devnums/pyheartlib/commit/1ea3f4e1bd3560010d70872bfb38f038266a953f))
-
-* fix: update ([`3b82ccf`](https://github.com/devnums/pyheartlib/commit/3b82ccf6152358a833121b6e61a4847657ce5578))
-
-
-## v1.0.5 (2023-11-10)
-
-### Fix
-
-* fix: update ([`81e8a67`](https://github.com/devnums/pyheartlib/commit/81e8a67a44fccbcb21398f35aa370a8ea95cbfd4))
-
-* fix: update ([`715796a`](https://github.com/devnums/pyheartlib/commit/715796a275011b0effd648808e57e461ea6115b1))
-
-
-## v1.0.4 (2023-11-10)
-
-### Fix
-
-* fix: update ([`6edf6b0`](https://github.com/devnums/pyheartlib/commit/6edf6b0bee2ee57180c18bd79621780eb3494dba))
-
-* fix: add fields ([`c0806ef`](https://github.com/devnums/pyheartlib/commit/c0806ef8f9e0c6a98c198c399d476b8d143aa7f3))
-
-
-## v1.0.3 (2023-11-10)
-
-### Build
-
-* build: update ([`b17d089`](https://github.com/devnums/pyheartlib/commit/b17d089702c6d466486ffe12a1d398fc8cb0f83d))
-
-### Ci
-
-* ci: update ([`d2e8774`](https://github.com/devnums/pyheartlib/commit/d2e8774d79dc0d56720429d7cac221bec3040111))
-
-### Fix
-
-* fix: update ([`959b12d`](https://github.com/devnums/pyheartlib/commit/959b12d9975eccea721b701ed2036d896bb4ccd2))
-
-### Unknown
-
-* Merge branch &#39;main&#39; of https://github.com/devnums/pyheartlib ([`adbc4a8`](https://github.com/devnums/pyheartlib/commit/adbc4a8f19695c073947a11ba59b11a619441cc3))
 
 
 ## v1.0.2 (2023-11-10)
@@ -283,297 +221,6 @@
 
 * chore: change to new license ([`4786892`](https://github.com/devnums/pyheartlib/commit/4786892c835afc8980448f84643e9272908e1194))
 
-
-## v1.0.0 (2023-06-17)
-
-### Breaking
-
-* build: update version
-
-BREAKING CHANGE: update for release. ([`6a89ca6`](https://github.com/devnums/pyheartlib/commit/6a89ca6ee2a436c3cfb99e3572b1cc7b3aa0253c))
-
-### Build
-
[…]
