### 2023-11-11 df1cf4a
1.0.7
Automatically generated by python-semantic-release
