+++ CHANGELOG.md
@@ -1,5 +1,94 @@
-# Changelog
+# CHANGELOG
 
+
+
+## v1.0.7 (2023-11-11)
+
+### Build
+
+* build: update version source ([`f41c5b2`](https://github.com/devnums/pyheartlib/commit/f41c5b27b32b250279d4fb789199f47b8a095587))
+
+### Chore
+
+* chore: edit file ([`d601914`](https://github.com/devnums/pyheartlib/commit/d601914ff38f7067bc7995e58dcb9b23c9b3f681))
+
+* chore: update file ([`b19849b`](https://github.com/devnums/pyheartlib/commit/b19849bade38d6adf95f1aa4c5dcddba797a5122))
+
+### Ci
+
+* ci: fix style ([`47bcdcf`](https://github.com/devnums/pyheartlib/commit/47bcdcfc9dbca865feb9d9d6bf27f17a8518fdeb))
+
+### Documentation
+
+* docs: update images ([`b2e31b4`](https://github.com/devnums/pyheartlib/commit/b2e31b4bb43adcd6ae32c2822c24807a4f5a1cc1))
+
+* docs: add new dep to the requirements ([`f810572`](https://github.com/devnums/pyheartlib/commit/f81057263bd47788afbe2d7f1a5eab89f4509260))
+
+### Fix
+
+* fix: bump version! ([`23d6029`](https://github.com/devnums/pyheartlib/commit/23d60294e5374b0cfe9f5d1cd5bbcc9eb851d7a1))
+
+
+## v1.0.6 (2023-11-10)
+
+### Build
+
+* build: update field ([`336c3da`](https://github.com/devnums/pyheartlib/commit/336c3da44fa74a9fb58b99c880d1b1c7a0fbbf2e))
+
+* build: add a new dep ([`6ae7c8f`](https://github.com/devnums/pyheartlib/commit/6ae7c8fe597a85fa4e124044b6a88c5201064fff))
+
+### Ci
+
+* ci: update workflow ([`76d2814`](https://github.com/devnums/pyheartlib/commit/76d28143c516b55b10994619a8ac8daa0dfeaee3))
+
+### Documentation
+
+* docs: add requirement ([`43b86ff`](https://github.com/devnums/pyheartlib/commit/43b86ffbc0589144d5f3793fb0e313ed8b08d936))
+
+* docs: fix ([`a821e6f`](https://github.com/devnums/pyheartlib/commit/a821e6f140df57ed0ebe569454d81189d933cbf5))
+
+### Fix
+
+* fix: fix bug ([`1ea3f4e`](https://github.com/devnums/pyheartlib/commit/1ea3f4e1bd3560010d70872bfb38f038266a953f))
+
+* fix: update ([`3b82ccf`](https://github.com/devnums/pyheartlib/commit/3b82ccf6152358a833121b6e61a4847657ce5578))
+
+
+## v1.0.5 (2023-11-10)
+
+### Fix
+
+* fix: update ([`81e8a67`](https://github.com/devnums/pyheartlib/commit/81e8a67a44fccbcb21398f35aa370a8ea95cbfd4))
+
+* fix: update ([`715796a`](https://github.com/devnums/pyheartlib/commit/715796a275011b0effd648808e57e461ea6115b1))
+
+
+## v1.0.4 (2023-11-10)
+
+### Fix
+
+* fix: update ([`6edf6b0`](https://github.com/devnums/pyheartlib/commit/6edf6b0bee2ee57180c18bd79621780eb3494dba))
+
+* fix: add fields ([`c0806ef`](https://github.com/devnums/pyheartlib/commit/c0806ef8f9e0c6a98c198c399d476b8d143aa7f3))
+
+
+## v1.0.3 (2023-11-10)
+
+### Build
+
[…]
