+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pyheartlib"
-version = "1.0.7"
+version = "1.2.0"
 description = "A Python package for processing electrocardiogram signals"
 authors = ["devnums <devnums.code@gmail.com>"]
 license = "AGPL-3.0-only"
@@ -40,8 +40,8 @@ sphinx-togglebutton = ">=0.3.2"
 tomli = ">=2.0.1"
 
 [tool.semantic_release]
-#version_variable = "pyproject.toml:tool.poetry.version"
-version_toml = ["pyproject.toml:tool.poetry.version"]
+version_variable =  ["pyproject.toml:tool.poetry.version"]
+#version_toml = ["pyproject.toml:tool.poetry.version"]
 branch = "main"
 changelog_file = "CHANGELOG.md"
 build_command = "pip install poetry && poetry build"
