### 2023-11-11 9ddbb67
1.0.8
Automatically generated by python-semantic-release
