+++ pyproject.toml
@@ -41,7 +41,7 @@ tomli = ">=2.0.1"
 
 [tool.semantic_release]
 version_variable =  ["pyproject.toml:tool.poetry.version"]
-#version_toml = ["pyproject.toml:tool.poetry.version"]
+version_toml = ["pyproject.toml:tool.poetry.version"]
 branch = "main"
 changelog_file = "CHANGELOG.md"
 build_command = "pip install poetry && poetry build"
