+++ docs/introduction.md
@@ -4,6 +4,8 @@
 
 `Pyheartlib` is a Python package for processing electrocardiogram (ECG) recordings. This software facilitates working with signals for tasks such as heartbeat detection, heartbeat classification, and arrhythmia classification. Utilizing it, researchers can focus on these tasks without the burden of designing data processing modules. The package transforms original data into processed signal excerpts and their computed features in order to be used for training various machine learning models including advanced deep learning models, which can be trained by taking advantage of Keras and Tensorflow libraries.
 
+Recent advances in deep learning have made it possible to accomplish tasks previously unattainable through the implementation of conventional algorithms. The capabilities of conventional approaches for analyzing ECG signals are limited. Therefore, it is imperative to develop data-driven models capable of automatically identifying patterns from vast heterogeneous data. `Pyheartlib` supports this goal by providing data processing modules to assist researchers with the implementation of their own models for various tasks.
+
 ## Input Data
 
 `Pyheartlib` utilizes the WFDB Python package[^1] to read annotated ECG records, therefore it supports the WFDB format. Recordings with different formats can be converted using the WFDB Python package. In the WFDB format, each record contains a set of files for the raw signal and its annotations. When the WFDB Python package reads a record, it provides the raw signal as an array, and its annotations and their locations on the raw signal as separate lists. For example, a signal with 500,000 samples may have around 2000 annotations for the R-peak labels (the heartbeat type placed at the peak of the R-wave) and the rhythm types. The figure presented below depicts a portion of a lengthy signal. Further details regarding the WFDB format can be obtained from [this link](https://github.com/MIT-LCP/wfdb-python). After reading a record, `pyheartlib` makes a data structure for the signal and its annotations. This data structure contains the heartbeat annotations (R-peak labels), rhythm annotations, annotation locations on the raw signal, and the raw signal itself.
