### 2023-11-13 devnums/pyheartlib#12 @devnums
The package was successfully tested on both Ubuntu and macOS. That error might be caused by ECGSequence, which uses TensorFlow.

Per your suggestion, I have included requirements sections in the [README](https://github.com/devnums/pyheartlib/blob/main/README.md) file and documentation to mention that it works on x86-64 processors. Please visit https://pyheartlib.readthedocs.io/en/latest/getting-started.html#requirements
