### 2023-11-13 devnums/pyheartlib#13 @devnums
As per your suggestion, I have included additional information for each step. The example is now available as a Jupyter Notebook and can also be run in the Google Colab (Colab link is included at the top of the file). Please let me know if I have added enough information.
https://pyheartlib.readthedocs.io/en/latest/examples/rpeak_detection.html
