### 2023-11-13 devnums/pyheartlib#14 @devnums
The documentation now includes a page that explains all the examples.
https://pyheartlib.readthedocs.io/en/latest/examples/examples.html

Both examples 3 and 4 show how the `RpeakData` class can be used to create datasets. However, in Example 4, there are steps to use the datasets for training a deep learning model. In both examples, the `RpeakData` class is used, but the output datasets are different. Example 4 demonstrates a specific use case for it.
