### 2023-11-13 #5792 @devnums
Thanks for taking the time to review this. Your valuable comments and suggestions greatly improved the quality of the documentation.

I have modified the documentation to include the suggestions you offered in the GitHub issues.

Furthermore, in order to address your points, I have made the following changes to the documentation.

1. Some plots and additional explanation regarding the input and output datasets have been added per your suggestion. Please refer to the introduction section of the documentation.
[Introduction](https://pyheartlib.readthedocs.io/en/latest/introduction.html#input-data)

2. Additional details regarding the classes have been provided, along with a figure depicting waveforms and features. Please refer to [this part](https://pyheartlib.readthedocs.io/en/latest/introduction.html#class-beatdata) of the introduction.

3. I have included a figure that shows a signal excerpt. It also includes a diagram illustrating excerpt classification. Furthermore, metadata and its benefit have been explained. Please refer to [this part](https://pyheartlib.readthedocs.io/en/latest/introduction.html#class-rhythmdata).

4. In the introduction, I have added the relevant parameters for the main classes, including RhythmData and ECGSequence classes. [link](https://pyheartlib.readthedocs.io/en/latest/introduction.html#class-rhythmdata)
Additional details have been added to the [second example](https://pyheartlib.readthedocs.io/en/latest/examples/arrhythmia.html#generate-data-samples), which illustrates how to use the RhythmData and ECGSequence classes.
Please note that I have changed the name of ArrhythmiaData to RhythmData. However, the former name will still work.

5. Additional details have been added to [example-3](https://pyheartlib.readthedocs.io/en/latest/examples/rpeak.html) and [example-4](https://pyheartlib.readthedocs.io/en/latest/examples/rpeak_detection.html) (R-peak detection), which demonstrate the use of the [RpeakData class](https://pyheartlib.readthedocs.io/en/latest/introduction.html#class-rpeakdata).

Additionally, a [Getting Started](https://pyheartlib.readthedocs.io/en/latest/getting-started.html) section has been added to help users with different use cases.

Thank you again for your insightful suggestions, I hope I have implemented all of them. Please let me know if there is anything I missed or if there is something I can improve.
