### 2023-11-23 e6346cb
test: fix for multichannel support
