### 2023-11-23 aa05103
test: fix multichannel support
