+++ src/pyheartlib/beat_info.py
@@ -16,8 +16,6 @@ import numpy as np
 from scipy import stats
 from scipy.fft import rfft, rfftfreq  # noqa: F401
 
-from pyheartlib.extra.pqrst import PQRST
-
 
 class BeatInfo:
     """
@@ -44,11 +42,11 @@ class BeatInfo:
     in_ms : bool, optional
         Whether to calculate RR-intervals in time (miliseconds) or samples,
         by default True
-    whole_waveform : list
-        Whole waveform, that is not trimmed.
-    bwaveform : list
+    input_waveform : numpy.ndarray
+        Input waveform, that is not trimmed (len_waveform, #channels).
+    bwaveform : numpy.ndarray
         Segmented beat waveform. Selected as a segment of
-        the whole waveform.
+        the input waveform (len_waveform, #channels).
     start_idx: int
         The index of the first sample (onset) of the waveform
         on the original signal.
@@ -79,8 +77,8 @@ class BeatInfo:
         data : dict
             A dict containing data about the beat with keys:
 
-            'waveform' : list
-                Waveform of the beat.
+            'waveform' : numpy.ndarray
+                Waveform (len_waveform, #channels).
             'rpeak_locs' : list
                 A list containing locations of rpeaks.
             'rec_id' : int
@@ -96,18 +94,16 @@ class BeatInfo:
         self.data = data
         self.label = data["label"]
         self.rpeaks = data["rpeak_locs"]
-        self.whole_waveform = data["waveform"]  # Whole waveform
+        self.input_waveform = data["waveform"]  # Input waveform
         self.rri = self.get_rris(in_ms=self.in_ms)
         self.rri_smpl = self.get_rris(in_ms=False)
         self.sdrri = self.get_sdrri()
-        self.bwaveform, self.wfpts = self.get_beat_waveform()  # beat waveform
+        (
+            self.bwaveform,
+            self.wfpts,
+        ) = self.get_beat_waveform()  # Trimmed waveform
         self.avail_features = self.available_features()
         self.features = self.compute_features()
-        if self.bwaveform is not None:
-            self.pqrst_dict = self.pqrst()
-        if self.label == "V":
-            pass
-            # self.__plot_wf() #plot for debug
 
     def available_features(self):
         """Returns available features that can be computed.
@@ -164,11 +160,24 @@ class BeatInfo:
         for f in features_names:
             ret = getattr(self, f)()
             # if the feature function returns one value
-            if not isinstance(ret, dict):
+            if isinstance(ret, (int, float)):
                 feature_dict[f] = ret
             # if the feature function returns multiple values as a dict
             elif isinstance(ret, dict):
-                feature_dict.update(ret)
+                try:
+                    keys = list(ret.keys())
+                    float(ret[keys[0]])
+                    for key, val in ret.items():
+                        feature_dict[key] = val
+                except TypeError:
+                    for key, val in ret.items():
+                        feature_dict[key] = tuple(val)
+            # for channels
+            elif isinstance(ret, tuple) or isinstance(ret, np.ndarray):
+                ret = tuple(ret)
+                for chnl, val in enumerate(ret):
+                    key = f"{f}(CH{chnl+1})"
+                    feature_dict[key] = val
         return feature_dict
 
     def get_beat_waveform(self, win=[-0.35, 0.65]):
[…]
