### 2023-11-25 8a49220
test: fix multichannel support
