+++ src/pyheartlib/beat_info.py
@@ -63,6 +63,37 @@ class BeatInfo:
     features : dict
         Dictionary containing computed features. Keys are feature names
         and values are features values.
+
+    Notes
+    -----
+    `BeatInfo` includes some predefined features, however custom features
+    can also be defined. Each custom feature definition must adhere to this
+    syntax and its name must start with *F_*:
+
+    >>> def F_new_feature(self):
+    >>>    return return_result
+
+    The `return_result` must be one of the following:
+
+    1) A real number.
+
+    2) A dictionary such as {"Feature_1": value_1, "Feature_2": value_2}.
+    Each value can be a real number. Alternatively, each value can be a
+    one-dimensional array, list, or tuple. In this case, their elements
+    must correspond to the channels.
+
+    3) Tuple or one-dimensional NumPy array. Their elements must correspond
+    to the channels. For example, the return value of `F_feat_new()` as a
+    tuple *(element1, element2)* will produce `F_feat_new(CH1)`
+    for *element1* and `F_feat_new(CH2)` for *element2*. The order of
+    channels is determined by the *CHANNEL* field in the *config.yaml* file.
+
+    4) A list. An output as a list will be a list.
+
+    The custom features can be added to the beatinfo object by using the
+    `add_features()` method, the list of all available features can be
+    obtained using the `available_features()` method, and to select desired
+    features for computation the `select_features()` method can be used.
     """
 
     def __init__(self, beat_loc=10, fs=360, in_ms=True):
@@ -172,12 +203,14 @@ class BeatInfo:
                 except TypeError:
                     for key, val in ret.items():
                         feature_dict[key] = tuple(val)
-            # for channels
+            # For channels
             elif isinstance(ret, tuple) or isinstance(ret, np.ndarray):
                 ret = tuple(ret)
                 for chnl, val in enumerate(ret):
                     key = f"{f}(CH{chnl+1})"
                     feature_dict[key] = val
+            elif isinstance(ret, list):
+                feature_dict[f] = ret
         return feature_dict
 
     def get_beat_waveform(self, win=[-0.35, 0.65]):
