### 2023-11-29 aeb2ccd
docs: update and add new info
