### 2023-12-01 e4e2bb0
fix: add new methods, docstrings, clean some parts
