+++ src/pyheartlib/data_rhythm.py
@@ -13,7 +13,6 @@ import os
 
 os.environ["TF_CPP_MIN_LOG_LEVEL"] = "3"
 import math  # noqa: E402
-from warnings import warn  # noqa: E402
 
 import numpy as np  # noqa: E402
 from tensorflow.keras.utils import Sequence  # noqa: E402
@@ -210,12 +209,15 @@ class ECGSequence(Sequence):
         feature computation, by default 36
     shuffle : bool, optional
         If True, after each epoch the samples are shuffled, by default True
+    rri_output: bool, optional
+        Whether to return RR-intervals and their features.
+        If False, returns only waveforms.
     rri_length : int, optional
         Length of the output RR-intervals list.
         It is zero-padded on the right side, by default 150
 
-    Example
-    -------
+    Examples
+    --------
     >>> from pyheartlib.data_rhythm import ECGSequence
     >>> trainseq = ECGSequence(
     >>>    annotated_records,
@@ -225,9 +227,32 @@ class ECGSequence(Sequence):
     >>>    raw=True,
     >>>    interval=36,
     >>>    shuffle=False,
+    >>>    rri_output=True,
     >>>    rri_length=25
     >>> )
 
+    Notes
+    -----
+    Returns a tuple containing two elements when its object is utilized in
+    this way: ECGSequence_object[BatchNo].
+
+    The first element (Batch_x) contains data samples and the
+    second one (Batch_y) their associated annotation.
+
+    If `rri_output` is True, Batch_x is a list of NumPy arrays of
+    Batch_wave, Batch_rri, Batch_rri_feat.
+
+    Batch_wave contains signal excerpts or their features, Batch_rri
+    contains RR-intervals, and Batch_rri_feat contains RR-interval features.
+
+    If `rri_output` is False, Batch_x contains Batch_wave only.
+
+    If `raw` is False, Batch_wave has the shape of
+    (Batch size, Number of channels, Number of sub-segments, Number
+    of features), otherwise, it has the shape of
+    (Batch size, Number of channels, Length of excerpt).
+
+    Batch_y has the shape of (Batch size, ).
     """
 
     def __init__(
@@ -261,20 +286,23 @@ class ECGSequence(Sequence):
         Returns
         -------
         tuple
-            Contains batch_x, batch_y, where batch_x is a list of
-            numpy arrays of batch_seq, batch_rri, batch_rri_feat.
+            Contains batch_x and batch_y.
 
-            If raw is False, batch_seq has the shape of
+            If `rri_output` is True, batch_x is a list of Numpy arrays of
+            batch_wave, batch_rri, batch_rri_feat. If `rri_output` is
+            False, batch_x contains batch_wave only.
+
+            If `raw` is False, batch_wave has the shape of
             (batch_size, #channels, #sub-segments, #features), otherwise,
-            it has the shape of (batch_size, #channels, seq_len).
+            it has the shape of (batch_size, #channels, wave_len).
 
             batch_y has the shape of (batch_size, 1).
         """
         batch_samples = self.samples_info[
             idx * self.batch_size : (idx + 1) * self.batch_size
         ]
-        batch_seq = []
-        batch_label = []
+        batch_wave = []
+        batch_annotation = []
         batch_rri = []
         for sample in batch_samples:
             # eg sample:[10,500,800,'AFIB'] ::: [rec,start,end,ann]
@@ -285,7 +313,7 @@ class ECGSequence(Sequence):
             anno = ann
             if self.class_labels is not None:
                 anno = self.get_integer(ann)
-            batch_label.append(anno)
+            batch_annotation.append(anno)
             seq = self.data[rec_id]["signal"][start:end]
[…]
