### 2023-12-01 1dfe203
fix: improve record reading
