+++ src/pyheartlib/features.py
@@ -82,7 +82,7 @@ def get_stat_features(data, features="all"):
     return features_arr
 
 
-def get_hrv_features(rri, features="all"):
+def get_hrv_features(rri=None, features="all", only_names=False):
     """
     Computes hrv features for the input samples.
 
@@ -116,91 +116,97 @@ def get_hrv_features(rri, features="all"):
     else:
         flist = features  # features list
 
-    num_samples = rri.shape[0]
-    num_features = len(flist)
-    features_arr = np.ndarray((num_samples, num_features), dtype=np.float32)
+    if not only_names:
+        num_samples = rri.shape[0]
+        num_features = len(flist)
+        features_arr = np.ndarray(
+            (num_samples, num_features), dtype=np.float32
+        )
 
-    # successive RR interval differences
-    sd = np.diff(rri, axis=1)
+        # successive RR interval differences
+        sd = np.diff(rri, axis=1)
 
-    # calculate meanrr
-    if bool({"meanrr", "nsdrr", "nrmssd"} & set(flist)):
-        meanrr = np.mean(rri, axis=1)
+        # calculate meanrr
+        if bool({"meanrr", "nsdrr", "nrmssd"} & set(flist)):
+            meanrr = np.mean(rri, axis=1)
 
-    # calculate meanrr
-    if bool({"sdrr", "nsdrr"} & set(flist)):
-        sdrr = np.std(rri, axis=1)
+        # calculate meanrr
+        if bool({"sdrr", "nsdrr"} & set(flist)):
+            sdrr = np.std(rri, axis=1)
 
-    # calculate rmssd
-    if bool({"rmssd", "nrmssd"} & set(flist)):
-        rmssd = np.sqrt(np.mean(sd**2, axis=1))
+        # calculate rmssd
+        if bool({"rmssd", "nrmssd"} & set(flist)):
+            rmssd = np.sqrt(np.mean(sd**2, axis=1))
 
-    # calculate hr
-    if bool({"meanhr", "maxhr", "minhr", "medianhr", "sdhr"} & set(flist)):
-        hr = 60000 / rri
+        # calculate hr
+        if bool({"meanhr", "maxhr", "minhr", "medianhr", "sdhr"} & set(flist)):
+            hr = 60000 / rri
 
-    if "meanrr" in flist:
-        ix = flist.index("meanrr")
-        features_arr[:, ix] = meanrr
+        if "meanrr" in flist:
+            ix = flist.index("meanrr")
+            features_arr[:, ix] = meanrr
 
-    if "sdrr" in flist:
-        ix = flist.index("sdrr")
-        features_arr[:, ix] = sdrr
+        if "sdrr" in flist:
+            ix = flist.index("sdrr")
+            features_arr[:, ix] = sdrr
 
-    if "medianrr" in flist:
-        ix = flist.index("medianrr")
-        features_arr[:, ix] = np.ma.median(rri, axis=1)
+        if "medianrr" in flist:
+            ix = flist.index("medianrr")
+            features_arr[:, ix] = np.ma.median(rri, axis=1)
 
-    if "rangerr" in flist:
-        ix = flist.index("rangerr")
-        features_arr[:, ix] = np.amax(rri, axis=1) - np.amin(rri, axis=1)
+        if "rangerr" in flist:
+            ix = flist.index("rangerr")
+            features_arr[:, ix] = np.amax(rri, axis=1) - np.amin(rri, axis=1)
 
-    if "nsdrr" in flist:
-        ix = flist.index("nsdrr")
-        features_arr[:, ix] = sdrr / meanrr
+        if "nsdrr" in flist:
[…]
