+++ tests/dummy_data/config.yaml
@@ -12,10 +12,12 @@
 DATA_DIR: ""
 SAMPLING_RATE: 360
 CHANNEL: ['I', 'MLII']
-RECORDS: [100, 101, 103, 105, 106, 107, 108, 109, 111, 112, 113, 114, 115, 116, 117, 118, 119, 121, 122, 123, 124, 200, 201, 202, 203, 205, 207, 208, 209, 210, 212, 213, 214, 215, 217, 219, 220, 221, 222, 223, 228, 230, 231, 232, 233, 234]
-DS1: [101, 106, 108, 109, 112, 114, 115, 116, 118, 119, 122, 124, 201, 203, 205, 207, 208, 209, 215, 220, 223, 230]
-DS2: [100, 103, 105, 111, 113, 117, 121, 123, 200, 202, 210, 212, 213, 214, 219, 221, 222, 228, 231, 232, 233, 234]
 BEAT_TYPES: ['N', 'L', 'R', 'j', 'e', 'V', 'E', 'A', 'S', 'a', 'J', 'F', 'f', '/', 'Q']
 RHYTHM_TYPES: ['(AB', '(AFIB', '(AFL', '(B', '(BII', '(IVR', '(N', '(NOD', '(P', '(PREX', '(SBR', '(SVTA', '(T', '(VFL', '(VT']
+
+# Optional fields
+RECORDS: [100, 101, 103, 105, 106, 107, 108, 109, 111, 112, 113, 114, 115, 116, 117, 118, 119, 121, 122, 124, 200, 201, 202, 203, 205, 207, 208, 209, 210, 212, 213, 214, 215, 217, 219, 220, 221, 222, 223, 228, 230, 231, 232, 233, 234]
+DS1: [101, 106, 108, 109, 112, 115, 116, 118, 119, 122, 201, 203, 205, 207, 208, 209, 215, 220, 223, 230]
+DS2: [105, 111, 113, 121, 200, 202, 210, 212, 213, 214, 219, 221, 222, 228, 231, 232, 233, 234]
 INT_MAP_DICT: {'N': 0, 'L': 1, 'R': 2, 'V': 3, '/': 4, 'A': 5, 'F': 6, 'f': 7, 'j': 8, 'a': 9, 'E': 10, 'J': 11, 'e': 12, 'S': 13, 'Q': 14}
 MAP_AAMI: {'N': 'N', 'L': 'N', 'R': 'N', 'j': 'N', 'e': 'N', 'V': 'V', 'E': 'V', 'A': 'S', 'S': 'S', 'a': 'S', 'J': 'S', 'F': 'F', 'f': 'Q', '/': 'Q', 'Q': 'Q'}
