### 2023-12-01 29b7eb3
chore: update and rerun examples
