+++ examples/model/result.txt
@@ -4,16 +4,16 @@ Without considering adjacency of ground truth and prediction
 
               precision    recall  f1-score   support
 
-           0       1.00      1.00      1.00   5790290
-           1       0.93      0.91      0.92    123310
+           0       1.00      1.00      1.00   4733482
+           1       0.95      0.94      0.94    104918
 
-    accuracy                           1.00   5913600
-   macro avg       0.96      0.96      0.96   5913600
-weighted avg       1.00      1.00      1.00   5913600
+    accuracy                           1.00   4838400
+   macro avg       0.97      0.97      0.97   4838400
+weighted avg       1.00      1.00      1.00   4838400
 
 Confusion Matrix
-[[5781231    9059]
- [  10684  112626]]
+[[4728110    5372]
+ [   6540   98378]]
 
 
 ************************************************************
@@ -23,13 +23,13 @@ tolerance for the adjacency of ground truth and prediction
 
               precision    recall  f1-score   support
 
-           0       1.00      1.00      1.00   5790290
-           1       1.00      0.98      0.99    123310
+           0       1.00      1.00      1.00   4733482
+           1       1.00      0.98      0.99    104918
 
-    accuracy                           1.00   5913600
-   macro avg       1.00      0.99      0.99   5913600
-weighted avg       1.00      1.00      1.00   5913600
+    accuracy                           1.00   4838400
+   macro avg       1.00      0.99      0.99   4838400
+weighted avg       1.00      1.00      1.00   4838400
 
 Confusion Matrix
-[[5790081     209]
- [   2266  121044]]
+[[4733238     244]
+ [   2559  102359]]
