### 2023-12-01 82726ef
1.21.0
Automatically generated by python-semantic-release
