### 2024-01-12 devnums/pyheartlib#15 @Bsingstad
Why do you do the beat, arrhytmia and rhythm processing
As I understand your paper and your work, your main contribution is your deep learning-based R-peak detection. Therefore, as opposed to @cudmore, I do understand the relationship between the R-peak dataset and the model, as the generation of R-peak dataset is required to train the model. However, I don't understand how the beat, arrhytmia and rhythm processing helps the user to train/test the R-peak detection model?
