### 2024-01-12 devnums/pyheartlib#16 @Bsingstad
Jupyter Notebook - Duplicate code in markdown
In your notebooks, you often show the exact same code in the code cell and in markdown. T this is not common practice and leads to unnecessarily much content in your notebook.
