### 2024-01-12 devnums/pyheartlib#17 @Bsingstad
Test model on separate dataset
As far as I can see you are training, testing and validating your model on the same dataset; MIT BIH Arrhythmia Database. To validate that this model also works on other datasets (possibly using other hardware, measurement set up and so on) would be a huge strange for this software package in my opinion.
