### 2024-01-22 08b3729
test: add test for custom processors
