+++ src/pyheartlib/data_beat.py
@@ -50,6 +50,10 @@ class BeatData(Data):
         Parameter of the low pass-filter, by default 15
     progress_bar : bool, optional
         Whether to display a progress bar, by default True
+    processors : list, optional
+        Ordered list of functions' names for preprocessing the raw signals.
+        Each function takes a one-dimensional NumPy array as its input and
+        returns an array of the same length.
 
     Examples
     --------
@@ -81,6 +85,7 @@ class BeatData(Data):
         cutoff=45,
         order=15,
         progress_bar=True,
+        **kwargs,
     ):
         super().__init__(
             base_path,
@@ -88,6 +93,7 @@ class BeatData(Data):
             lowpass,
             cutoff,
             order,
+            **kwargs,
         )
         self.syms = self.config["BEAT_TYPES"]
         self.win = win
