### 2024-01-27 a1dda73
1.22.0
Automatically generated by python-semantic-release
