+++ paper/paper.md
@@ -27,19 +27,19 @@ Electrocardiogram (ECG) signals represent the electrical activity of the heart a
 
 Recent advances in deep learning have made it possible to accomplish tasks previously unattainable through the implementation of conventional algorithms. The capabilities of conventional approaches for analyzing ECG signals are limited. Therefore, it is imperative to develop data-driven models capable of automatically identifying patterns from vast heterogeneous data. `Pyheartlib` supports this goal and, unlike popular packages such as NeuroKit2 [@NeuroKit2_2021], does not implement pre-defined algorithms. It instead provides data processing modules to assist researchers with the implementation of their own models for various tasks. For example, SleepECG [@SleepECG_2023] and Neurokit2 [@NeuroKit2_2021] have implemented several conventional methods, such as the Pan–Tompkins [@PanTompkins1985] and Hamilton [@Hamilton2002] algorithms, for the R-peak detection task. However, the same task can be accomplished through deep learning. Various models can be trained by utilizing the readily available data processing modules in `Pyheartlib`. An example model is provided with this package that performs this task. `Pyheartlib` comprises several dedicated classes for different use cases.
 
-For the heartbeat classification task, which typically requires segmented and annotated heartbeat waveforms, the software processes the input data and delivers datasets comprising waveforms and features. Features can be computed from waveform and RR-intervals \autoref{fig:fig1}.
+For the heartbeat classification task, which typically requires segmented and annotated heartbeat waveforms, the software processes the input data and delivers datasets comprising waveforms and features. Features can be computed from waveform and RR-intervals (see \autoref{fig:fig1}).
 
 ![Pyheartlib processes the provided ECG records and generates a dataset comprising waveforms, computed features, and their corresponding annotations. \label{fig:fig1}](fig1.png){ width=98% }
 
-For the classification of signal excerpts, e.g., arrhythmia classification, the software is designed to store metadata about the excerpts in order to reduce memory usage significantly, especially in situations that each excerpt is being used only once during model training. By adjusting the parameters appropriately, the package provides a substantial quantity of data samples for training deep learning models. Moreover, it is feasible to incorporate RR-intervals in addition to waveforms and their computed features \autoref{fig:fig2}.
+For the classification of signal excerpts, e.g., arrhythmia classification (\autoref{fig:fig2}), the software is designed to store metadata about the excerpts in order to reduce memory usage significantly, especially in situations that each excerpt is being used only once during model training. By adjusting the parameters appropriately, the package provides a substantial quantity of data samples for training deep learning models. Moreover, it is feasible to incorporate RR-intervals in addition to waveforms and their computed features.
 
-![An excerpt, its onset, and offset on the original signal. The excerpt onset and offset are computed by a sliding window. \label{fig:fig2}](fig2.png){ width=98% }
+![An excerpt, its onset, and offset on the original signal. \label{fig:fig2}](fig2.png){ width=98% }
 
 Another use case is when each excerpt has to be divided into smaller sub-segments each with a specific label, e.g., R-peak detection. The package delivers data samples by storing metadata about the excerpts and providing lists of labels as annotations for the excerpts.
 
-`Pyheartlib` is easy to use and its documentation contains examples for different use cases. \autoref{fig:example} illustrates an example outcome of a deep learning model that was trained with the help of `Pyheartlib` and Keras [@chollet2015keras] to detect R-peaks.
+`Pyheartlib` is easy to use and its documentation contains examples for different use cases. \autoref{fig:fig3} illustrates an example outcome of a deep learning model that was trained with the help of `Pyheartlib` and Keras [@chollet2015keras] to detect R-peaks.
 
[…]
