### 2024-01-28 devnums/pyheartlib#15 @devnums
This software is useful for a wide range of applications, not only R-peak detection. R-peak detection is just an example use case.
