### 2024-01-28 devnums/pyheartlib#17 @devnums
This software can be used with other datasets and different machine learning models. It is only required that the datasets have a similar format and structure to the MIT-BIH Arrhythmia Database.
