### 2024-02-09 #5792 @Bsingstad
> > Thank you, @devnums, for developing an interesting Python package.
> > I have created 3 GitHub issues I would like to address:
> >
> > * [Why do you do the beat, arrhytmia and rhythm processing devnums/pyheartlib#15](https://github.com/devnums/pyheartlib/issues/15)
> > * [Jupyter Notebook - Duplicate code in markdown devnums/pyheartlib#16](https://github.com/devnums/pyheartlib/issues/16)
> > * [Test model on separate dataset devnums/pyheartlib#17](https://github.com/devnums/pyheartlib/issues/17)
> >
> > In addition, I have some other aspects of your work I also would like to address:
> >
> > 1. In this work you only use the MIT BIH Arrhythmia Database, containing ECGs from 47 subjects, both in training, validation and test. I would like to know more about how you partitioned the data? Do you split the data patient-wise or can the data from the same patient be in both the training, validation and test? This should at least be stated in the documentation.
> > 2. In the processing steps you are using WFDB package to import data and annotations, and I can see that you are using some familiar filtering techniques to remove baseline wander and noise. Can you explain if your work contribute with something novel on the processing side?
> > 3. As far as I understand your model only allow one ECG lead at a time, which makes sense for many application like wearbles, etc. Have you done any experiments assessing wich lead the model perform best at? Would assume the leads that usually have prominent peaks (like lead I, II, V4, V5 and V6), but it would be interesting to see such results anyway.
> > 4. It would be interesting to see comparison between the results of your work and previous work (i.e rule-based R-peak detection). Maybe there are some benchmark task you can apply your model to.
> >
> > In general, I think your paper is very short and this explains why important information about the dataset, your contribution in relation to previous contributions, implications, etc. are missing. However, the code you provided are fairly well organized and the aim of the software package, deep learning based R-peak detection, is very relevant and could potentially have a high impact in the field.
>
> Thank you @Bsingstad for taking the time to review this software package. I have answered the issues. Below, I provide answers to your points.
>
> 1. Users can choose which records to use for training, validation, and testing. They can be from multiple raw datasets. Each record has the data for one patient. When using the BeatData class, users can also create intra-patient datasets. Therefore, a record can be utilized in both training and testing datasets.
> 2. I have added functionality that will allow users to use their own custom preprocessing methods. Please see the link below.
>    https://pyheartlib.readthedocs.io/en/latest/getting-started.html#optional-preprocessing
> 3. The software supports multiple channels. The provided example only uses Lead II; however, users can design different models and use different channels or multiple channels at once.
> 4. That's a possibility for research in that specific area. Here, I have provided an example usage for this software.
>
> I appreciate you taking the time to read the paper and the documentation and provide feedback.

Thanks for you answers and I am happy to see the improvement of your manuscript
