### 2024-02-11 devnums/pyheartlib#17 @Bsingstad
Ok, then it would be nice of you to show this in the paper. From what you have showed we can't know if the model is overfitted to the specific dataset (MIT-BIH Arrhythmia Database).
