+++ paper/paper.bib
@@ -40,7 +40,7 @@
       Gow, B. and
       Moody, B.},
   year={2023},
-  journal={PhysioNet},
+  publisher={PhysioNet},
   doi={10.13026/9njx-6322}
 }
 
@@ -48,7 +48,7 @@
   title={WFDB Software Package (version 10.6.2)},
   author={Moody, G. and Pollard, T. and Moody, B.},
   year={2021},
-  journal={PhysioNet},
+  publisher={PhysioNet},
   doi={10.13026/zzpx-h016}
 }
 
@@ -151,7 +151,7 @@
   doi={10.1109/TBME.1985.325532}
 }
 
-@article{Hamilton2002,
+@inproceedings{Hamilton2002,
   author={Hamilton, P.},
   booktitle={Computers in Cardiology},
   title={Open source ECG analysis},
