### 2024-02-27 #5792 @britta-wstnr
Hi @devnums,

I double-checked the paper,  and I have two small suggestions that I will just paste here:
`line 8`: "for the diagnosis of various disease." --> "for the diagnosis of various disease**s**."
`line 37`: "especially in situations that each excerpt is being used" --> "especially in situations **where** each excerpt is being used"

Further, for completeness, I confirm that both Tensorflow and Keras (which `editorialbot` complains about regarding DOI) are cited as per the softwares' FAQs/requirements.

@devnums, if you agree please fix the typos above.
Could you then please go on with these points from our [post-review checklist](https://github.com/openjournals/joss-reviews/issues/5792#issuecomment-1946483148):

> - Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> - Archive the release on Zenodo/figshare/etc and post the DOI here.
> - Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> - Make sure that the license listed for the archive is the same as the software license.

Thanks! 🙏
