### 2024-03-01 #5792 @devnums
Hi @britta-wstnr,

Thank you for your suggestions. I have fixed the typos.

The version of the software is 1.22.0

I have archived the software on Zenodo. The DOI is: [10.5281/zenodo.10730468](https://doi.org/10.5281/zenodo.10730468)

Thanks again 🙏
