### 2024-03-08 #5792 @britta-wstnr
Okay! Everything in order here as far as I can see 🎉  @devnums I will hand this off to our EiC now for acceptance.

A big thank you to the reviewers @cudmore @LegrandNico and @Bsingstad - we deeply appreciate your review work for JOSS! 🙏 🌷

_A quick note to the EiC taking this over:_ I checked the two missing DOIs that the `editorialbot` flags and both Keras and Tensorflow are cited as the homepages for the softwares request as far as I can see.
