### 2024-03-08 #5792 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
